[roll] Roll fuchsia [fuchsia.sys2] Drop legacy support for StartResult
This removes `legacy=true` on the StartResult enum, which was removed in
I05811cdd39bacd6776e9fe95f65032af191f6d1a. It is no longer available in
any supported API level.

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Original-Bug: 42085274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996692
Original-Revision: 741f089b0573b088dba04223a8b65028e9c034cc
GitOrigin-RevId: 80d7f90d4963a00402d1fbdf25a03c09620018dd
Change-Id: I0334a0c296759bf5928dbee120390b8db555ee49