[roll] Roll fuchsia [kernel][pi] Fix a bookkeeping error in PI propagation.
Fix an error in PI bookkeeping. The short version of the issue is that
it was possible for the inherited dynamic scheduler parameters to be
incorrect if there was a situation where

1) Thread A with a (inheritable or not) fair profile blocked in an
   OwnedWaitQueue, W.
2) Thread B with an inheritable deadline profile blocked in W.

In this situation, when thread B blocked, the code was asking the
question "Is this the first thread to block in the queue?  If so I can
just copy the dynamic params from the blocking thread into the queue
and call it good."

This turns out to be the wrong question to ask.  The question should
instead be "Is this the first thread with defined dynamic scheduling
parameters to block in the queue?".

Failure to use the proper predicate can result in a 0 start time being
recorded as the inherited start time in the queue, which can lead to
nonsensical results (like a negative start time) later on when those
corrupt inherited dynamic parameters are combined with other dynamic
parameters down the line.

Fair threads, when they block, don't have a defined start or finish
time.  They became undefined when the thread blocked and the thread
will receive new parameters when it unblocks.  Deadline threads, on
the other hand, always have inheritable parameters, and keep their
dynamic parameters across blocking events.  We leverage this to fix
our predicate when maintaining the inherited dynamic scheduling
parameters for an owned wait queue.

Now, there are 4 cases to consider:

1) An OWQ has no deadline pressure before a thread blocks upstream of
   it, but acquires pressure from the upstream thread.  Only deadline
   threads can contribute dynamic parameters, and this is the first
   one, so we can just copy those parameters.
2) An OWQ had deadline pressure before, and has it after a thread
   blocked, and the pressure has changed.  We need to recompute the
   dynamic parameters for the OWQ by calling into the scheduler in
   order to apply the lag equation.
3) An OWQ had deadline pressure before and now it does not have any.
   Technically, nothing special needs to be done, but in a build with
   extra debug checks, we want to reset the dynamic parameters to
   their constant initial values to make things consistent, and enable
   some extra asserts (when we want to assert that the values are
   undefined)
4) The deadline pressure didn't change at all.  Nothing special needs
   to be done to the dynamic parameters.  Accumulating the static
   profile pressure is all we need to do.

So, fix the predicate and add a new test to make sure that things are
now behaving as they should.

Test: new unit test
Original-Fixed: 42182770
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994748
Original-Revision: 5ca7d3e8db382ab4cc7e627095deedf88038a5ae
GitOrigin-RevId: 9139c8252c4dd5b81b66a6cee2e6d79bf3846153
Change-Id: Ibe7fff6a34da15812393ea81551137a38e3519ee