[roll] Roll infra/recipes Use removeprefix/removesuffix
Now that recipes are on Python 3.11, we can use these new functions
introduced in Python 3.9, which simplify removing a prefix or suffix
from a string. They return the original the string if the original
string does not have the specified prefix/suffix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/941072
Original-Revision: 68fef9acb3e4465984047aa782af909fce6a5809
GitOrigin-RevId: e86ead6ad9095d70e99187885c49a5648721627e
Change-Id: I3e0d0357bdef810bb6482855f373a802ac0da31e