[roll] Roll fuchsia [fidl] Remove outdated tests and docs
Before we had versioning and an SDK it was extremely important to know
that we could do "soft transitions" when changing APIs whose
declarations and uses were split across different repositories. We built
detailed documentation, coupled with tests that validated that the
documentation was correct.

Now we have API levels so if we want to change the platform interface we
can do that simply and cleanly without breaking the build of any
repositories. The advice that used to be critical to keeping builds
green now causes errors.

It would be nice to have a replacement for this documentation and these
tests ready to go, but it's more important to stop telling people the
wrong thing.

Original-Bug: 42078310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995757
Original-Revision: c5a970d203c9d55d453d163e7fdced5ab973b3b8
GitOrigin-RevId: c0083ba7eb07a0986895a9fcd465dd37d8c77fd9
Change-Id: If1f7c4e434f9063930683993c6782094453320e2