[roll] Roll fuchsia [fuchsia.component.config] Add legacy=true
This adds legacy=true on the fuchsia.component.config library, which was
removed in API level 14, because it still exists in a supported API
level (12).

This library was removed in If825f4535a43d1c5fa59a60cd63aac5764100aa4
without legacy=true.

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Original-Bug: 42085274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997856
Original-Revision: a472dd2d2c063d0af87f9ee1c6cdb88deeb80361
GitOrigin-RevId: fce7400fc1fbea554c32a93781563ccf713f2b34
Change-Id: Idb5eb86c492c003ada316b3df589a95b2f87fb86