[roll] Roll infra/recipes [shac] Ignore changed files from different repos
If the Gerrit change being tested is in a different repository from the
repository being analyzed, don't consider any files affected. This
ensures we don't incorrectly add comments to files in the wrong
repository, which could happen if two repositories contained a file at
the same location.

I haven't seen any indication of this occurring yet, but it likely would
have come up eventually.

I also changed the `api.git.get_remotes()` function to return a list
instead of a generator to make it behave less surprisingly when used in
conjunction with `with api.context(cwd=...):`; when it returns a
generator, the underlying step doesn't run until the generator is
iterated over, which may be outside the intended context, as opposed to
when the function is called.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/998512
Original-Revision: 0ce45f31826a65c6fe6cda5157bceb83d9b249a8
GitOrigin-RevId: 0b4431cbb999c456ed5e5ac9b91952890f16dd31
Change-Id: Icb1d8dd61fb9cf3bd6a73a16487186b6d1bad553