[roll] Roll fuchsia [wlansoftmac] Add async Rx path tracing
In general, when tracing is off, each tracing call in the code is
"about the cost of a branch instruction" which is on the order of
nanoseconds. When tracing is on, it should cost a hashtable lookup
plus some small memcpy's which is on the order of <200 nanoseconds. In
other words, with tracing on or off, we would not expect tracing to
bottleneck data path performance until we're receiving at very high
data rates.

Test: Manually recorded a trace
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979817
Original-Revision: a9aa4747d13a996307b54ff6a38d68fe2cafddf7
GitOrigin-RevId: 91281ef88dfee7d03e666b55c3cedc0ab5d4e3ff
Change-Id: I762aec82267b71b8c4f01457d18765839bfe0e51