[roll] Roll fuchsia [fxfs] Use fixed-with ints for checksums
Stores checksums as an array of little-endian integers in a  Vec<u8>
with some custom zero-copy accessors to &[Checksum].  This bypasses
bincode's preference of using varint encoding for u64 types by default,
which for checksums is wasteful (9 bytes instead of 8) as well as quite
a bit slower to deserialize.

The zero-copy accessors require little-endianness, so I've asserted that
we are building Fxfs on that for now since it's unlikely to change.

I also found that the journal wasn't checking extents in the root store
for validity if they didn't have checksums, which I corrected (and
removed two fsck test cases that would catch this otherwise -- now the
journal simply ignores the malformed extents during replay).

Original-Bug: 323571978
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996156
Original-Revision: 9d07ad87ddc55830907462a7e436c741e1242de2
GitOrigin-RevId: bbdeade86a90dc33ec80a95721ee65cfb8f804a5
Change-Id: I55c85effd7140191a8ec56e8751fc64ffc335df5