[roll] Roll fuchsia [fastboot]: upload takes Reader
Previously upload function would take a slice of u8 to upload. While
this worked for small files this would require large files to be read
entirely into memory in order to be uploaded to the Device. This lead to
inneficient memory usage.

Now, the upload function takes an impl Reader and a size to inform the
Device how much to be uploaded.

Bytes are uploaded to the target in 4096 byte chunks.

This also allows for the on_progress option for the
UploadProgressListener to be called as preivously there was no progress.
Files were uploaded in their entirety or they failed.

Original-Bug: b/325501917
Original-Bug: b/326990810
Fix: b/325501917
Fix: b/326990810
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997672
Original-Revision: e69adb5e0cb7064331e9f53a4772829df07c1285
GitOrigin-RevId: 7d47ff54afdeaac2fc7de4b8c6f22aa8514e1a0b
Change-Id: I0e1b65702fd92a74e45722ff370ab598ad2c55b3