[roll] Roll fuchsia [driver_runtime] Migrate to Role Manager
Migrate the driver runtime and compatibility shim to using RoleManager.
Note that the fuchsia.scheduler.RoleManager capability was already
routed to all the relevant places in https://fxrev.dev/995272, this
change just modifies the code to use that capability.

A notable exception is the SPI code, which was committed after the
aforementioned CL, so this CL adds the necessary route.

This CL also removes an API level check that is no longer needed.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997934
Original-Revision: f50b8914f963e4e9aee2ae8a48877165c7960b62
GitOrigin-RevId: 02566d0c575dfaf4b1f3636071d2a8b9f6a808e1
Change-Id: If9fb23bbe1926cb1239d62d1064ddaacbc286ef0