[roll] Roll fuchsia [fidl][rust] Do not call get_unchecked on empty slices
This fixes a bug in FIDL Rust decoding that is failing assertions in the
canary builders that roll the Rust toolchain.

For certain types, such as vector<uint8>, we optimize decoding to a
memcpy. We were using `get_unchecked` to obtain the source pointer in
the buffer. This was incorrect for empty vectors because `get_unchecked`
returns a reference, not a pointer.

I made two changes. Either one on its own would fix the problem, but
they are both worthwhile:

1. Use slice.as_ptr().add(offset) instead of slice.get_unchecked(offset)
   before calling memcpy. This avoids creating a reference at all.

2. Assert that len > 0 in decoder.out_of_line_offset, as we already do
   in encoder.out_of_line_offset. This significantly simplifies that
   method's `unsafe` rationale comment. It only required adding len != 0
   checks to a few callers (most pass inline sizes that cannot be zero).

Test: fx test fidl_rust_conformance_tests
Original-Fixed: 327029914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997122
Original-Revision: 87026743e6678d403a210a47bf935c951271b8dc
GitOrigin-RevId: 8296960e70a97fedddf5135046a2d7b8e3da56e6
Change-Id: I7d3370eff7fd77568c2893872a66462473496cbb