[roll] Roll fuchsia [fuchsia.diagnostics] Add legacy=true
This adds legacy=true on the fuchsia.diagnostics/LogSettings method
RegisterInterest because it is removed=18, so it still exists in
supported API levels (the oldest supported level is currently 12).

This method was removed in I22d047d9981b33d63f2ebfe9629bebb8862efa74
without legacy=true.

There are two implementations of the LogSettings server. I changed them
to handle RegisterInterest as follows:

* src/developer/ffx/plugins/log/src/testing_utils.rs:
  Panic, since it's test code.
* src/diagnostics/archivist/src/logs/servers/log_settings.rs:
  Log a warning and close the channel.

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Original-Bug: 42085274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997718
Original-Revision: 27737417a5a4388eb63cfbd3b844959ae8bdf628
GitOrigin-RevId: a6c3d807ae9b920bf77c0860d3d8ffda39a594b2
Change-Id: Iebc704e80632fcdcba236d6fa857e405c2ae0f47