[roll] Roll fuchsia [starnix] Fix ptrace01 flake
Root cause was that a tracee notified its tracer of its exit before
it removed itself from the set of children.  If the tracer made two
calls to wait in a row, it could wait after the tracee sent the
notification, but before it told the tracer not to wait on it
anymore.

This change modifies the code so that a lock is held during the
notification / removal process.

Original-Bug: b/325437151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998772
Original-Revision: 61fbc24a86054de71986fe0206e9f2ee4c6cf390
GitOrigin-RevId: 6fd5a1d9714fec0611079302b138cd5e5c65c47d
Change-Id: I8d77204b1e30b576c6725976afa71802109dc088