[roll] Roll fuchsia [kernel][vm] Make AllocatedPagesLocked iterative
Change AllocatedPagesLocked so that it doesn't calculate allocated pages
recursively, as this has the potential to create an unbounded kernel
stack.

The const qualifier is removed from the function as it doesn't
make sense to propagate const to an enumerator at this time.

Original-Bug: 322261403
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995192
Original-Revision: 8002a9367add87b5c4dd90e549837310a3edad21
GitOrigin-RevId: 0ad7fdd1eeba9260b4735e11e835a8f95b834ea5
Change-Id: I5869b0912c93275068020f9745c78a20798f4f72