[roll] Roll fuchsia [versioning] Assert a valid API level has been specified
Previously, `availability` attributes would only be added if
`__Fuchsia_API_level__` is defined, but there was no guarantee in the
file that it is always defined when targeting Fuchsia or set to a valid
value. Add such a check and comments about why it must be defined.

Original-Bug: 42085580
Test: fx build && fx build final_fuchsia_idk
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/942773
Original-Revision: 56e9e30f8bb5ea3cca30df465914304413536f4e
GitOrigin-RevId: 114f2855a9433b9b158a18503221b3ae2ffd821b
Change-Id: I9360b5bae05580667bdeb8af1cc9362b0b302b84