[roll] Roll fuchsia [devfs][storage] Removing multiplexing in block-verity
As part of removing Devfs, this CL removes an
instance of fidl multiplexing, where a channel
is used for the the fuchsia_hardware_block::Block
an fuchsia_io::Node protocols.  This behavior is
hidden by the use of file descriptors.

Changed to now explicitly use typed ClientEnds.

The only use of the affected functionality of this driver is
in tests.

Original-Bug: 42063787

Testing: Updated relevant tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994963
Original-Revision: 12c05d0c23093283c022db516107c11bcc3e618c
GitOrigin-RevId: 8d290cbab5dd1171a789788fa64a13541bafc6d5
Change-Id: I6c0b782d6478f85484f560988da7c9eae45fc342