[roll] Roll fuchsia [component] Plumb escrow event from ELF runner to component_manager
This CL is pure plumbing and unit tests that the escrow is reliably
delivered.

c.f. escrow API [1], the `fuchsia.process.lifecycle/Lifecycle` API gains
a new `OnEscrow` event that ELF components can use to send away their
outgoing directory. The ELF runner will forward this to
component_manager using the `ComponentController` API.

The idea is that when a component is stopped, we'll reap whatever escrow
state that was stored in the Program object, and later put that to use
somewhere.

[1]: https://fxrev.dev/991412

Original-Bug: 319754472
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998792
Original-Revision: d5e24121f1e21029702347d63865201d65796cc9
GitOrigin-RevId: d812ca9a0af0be0ccba0d0b3dd6835cd6e234ab6
Change-Id: I8142d60f3dba7f3a68ee6aede54f973ec72eef0e