[roll] Roll fuchsia [fuchsia.component.decl] Add legacy=true
This adds legacy=true on fuchsia.component.decl declarations removed in
API levels 13, 14 and 15, because they still exist in a supported API
level (12).

These elements were removed in the these CLs without legacy=true:

* I52cf055ca1a471bd81a4013e023a4edae6cade3f
* I334599d6588e956598b70c1ef1fd3d8dcfe35207
* I46779dbbb3375d0b46e7a551e54f792cbbca6918

The motivation for this is to line up legacy=true with the current set
of supported API levels for the implementation of RFC-0232.

Original-Bug: 42085274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/997717
Original-Revision: c960a740fcacfada912108174515b7f681df167d
GitOrigin-RevId: abc89e7351481f770eadc9b93ac85ef3360fa4fb
Change-Id: I89fab0e97ee2769cdf56b4a1b8fa2fb4ac2665d8