[roll] Roll fuchsia [starnix] Implement first version of the ebpf verifier
The verifier ensures that an ebpf program cannot:
- Do any out of bound access
- Leak kernel address to userspace
- Use any uninitialized data

This first version is extremely conservative. A more subtle
implementation will require a more refined type system to be able to
keep track of comparison between instances.

This first version can also be improved from a performance point of
view. By keeping track of which state is correct at a given pc, it is
possible to cute computation when the same pc which a smaller state is
reached again.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999353
Original-Revision: 285414ff35e8ea90a3ae8b2bd3fb705d17bb1996
GitOrigin-RevId: 1261e8df3fee858bf6a31ef9b58805faea411689
Change-Id: Ieb4998588ec69197cb1a8dc51c91656f51d06c55