[roll] Roll fuchsia [ffx] Daemonless knock implementation
This change introduces FIDL-pipe, which is a simplified re-write of
`ffx/daemon/target/src/overnet/host_pipe.rs`, which is the backbone of
the daemonless discovery function. It has three streams, one being
input, one being output, and the last being stderr from the SSH command.

The daemonless knock function is implemented essentially as:

* Resolve target address (IP address or nodename are acceptable).
* SSH to target.
* Establish overnet node.
* Knock RCS.

An example of this being used is in the follow-up CL in which this is
being used for ffx emulator start.

Most functionality from host_pipe.rs (regarding ssh message parsing) is
retained and untouched, moved over to ffx/lib/ssh/parse.rs

And other code touched is either kept identical or slightly altered in
order for it to be usable in a library outside of the daemon.

Original-Bug: b/324322258

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982392
Original-Revision: 96b48b296e5b446733680f34116ccbf07b9c4a02
GitOrigin-RevId: dba79a86ce9d5db765650244f57a0b77f49f156d
Change-Id: I13952f513620707e4e2000979a818622f6935b54