[roll] Roll fuchsia [kernel][cmpctmalloc] Clean up build rules and header use
This cleans up the build rules for //zircon/kernel/lib/heap
and its subdirectories, which uncovered some missing deps
in other kernel code that is using `<lib/heap.h>`.

It narrows down which headers are used by what to what's actually
needed, which simplifies the picture for the host test build of
the cmpctmalloc code.  It also makes `<lib/heap.h>` not declare
the kernel functions when used outside the kernel, i.e. in the
host test program.  That test code really shouldn't interact with
that layer at all, but currently needs the HEAP_DEFAULT_ALIGNMENT
constant because the cmpctmalloc doesn't provide it though it
exists to match the hard-coded 8 in the cmpctmalloc code.

This all fixed a build regression when upstream libc++ headers
changed so that <stdlib.h> is sometimes included in places where
it wasn't before, and its declarations of the standard `malloc`
et al can conflict with the kernel-oriented declarations in
<lib/heap.h>.

Original-Fixed: 327668937
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000493
Original-Revision: 274fe68b065679f574f3cee405f8cf42a2f1b5d2
GitOrigin-RevId: 9eb65c0984f476cd1b081e0cb1903336d980eab3
Change-Id: I3662bd75b0c7819c8d2cc2fe57791a209965a920