[roll] Roll fuchsia [display][fidl] Rename Image.type to Image.tiling_type.
The name "type" is overly generic. The field's semantics is fairly
closely aligned with the Vulkan concept of image tiling. This CL renames
the field and all related constants.

This CL also renames the "simple" tiling type to "linear", for the same
reasons mentioned above.

In the bigger picture, this CL is a readability bandaid until we
redesign the mechanism for negotiating image properties.

Original-Bug: 316631158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959498
Original-Revision: 7b3faa89dd1016ffba94acb4555d2c7d38d53e66
GitOrigin-RevId: a76cf110dac9fbdc379d49c0b52a3f7ca4bb60fc
Change-Id: I818ab494adcd9c0f0de1b3a6e52ebe22cb561dff