[roll] Roll fuchsia [sdmmc] Merge the Banjo and FIDL unit tests.
sdmmc-block-device-fidl-test.cc is almost the same as
sdmmc-block-device-test.cc, except that the SDMMC protocol opreates
in FIDL in the prior and in Banjo in the latter. This change merges
the two tests into one in the form of a parameterized unit test.

The contents of sdmmc-block-device-fidl-test.cc is kept under the
name of sdmmc-block-device-test.cc. An additional few lines of changes
introduces the test parameterization (GetParam == true for FIDL, false
for Banjo).

The motivation for this change is to remove the burden of maintaining
what is essentially a same copy of the test.

Test: fx test sdmmc-block-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000253
Original-Revision: 7cec85f4b05e79a34d59c42dbdccb8ec75ceefbc
GitOrigin-RevId: 00fa71ccc1179743e55db5a10e9335946a5f7780
Change-Id: I70195655aa60a40fd1f7c7a2c8f9f0264dabc3e6