[roll] Roll fuchsia [rust][build] Temporarily allow unused_imports
Just for the sake of rolling the toolchain reducing the risk of races
for this high potential to race lint.

Expect this to be very short lived, as soon as the toolchain rolls this
lint will be re-enabled in a change that fixes any unused import that
was added in the meantime.

Original-Bug: 326330182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000279
Original-Revision: 6fcc2c70143152f87e4599db4a09ca081dedca92
GitOrigin-RevId: a0190fcf0d310b5d4c423f6bb153e030d7806fea
Change-Id: I3e6dd28661025c09d168eac2f65e22e50258be68