[roll] Roll fuchsia [mutex_pi_exerciser] Convert to using the zx_profile system calls
This CL changes the mutex_pi_exerciser to use the `zx_profile_create`
system call to create profile objects instead of going through the
ProfileProvider. This is part of the work to deprecate the
ProfileProvider.

The alternative of using the RoleManager (the ProfileProvider's
replacement) was considered, but that API is a bit too rigid for
the kinds of desk testing mutex_pi_exerciser runs.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001575
Original-Revision: f38b11a72aa9b63423f828a0688c62e17b723089
GitOrigin-RevId: 2d8339e122d0d8f8702276309a80c8e1a155415f
Change-Id: Iea164b6769f41c0c13b7a6e1725f9e3614f89dcb