[roll] Roll fuchsia [bt][core] refactor LowEnergyScanResult to its own class
The LowEnergyScanResult class was previously a struct with all public
members used to track the elements related to a scan result. A bit
awkwardly, the advertising data and scan response data related to the
scan result was stored in LowEnergyScanner::PendingScanResult.

This change updates LowEnergyScanResult to store the advertising data
and scan response data within itself. To present a better API to users,
and to prevent accidental modification, LowEnergyScanResult is now a
class with getters and setters instead of a struct with all public
members.

Since LowEnergyScanResult stores the advertising data and scan response
data itself, we can also modify the
LowEnergyScanner::Delegate::OnPeerFound API to simply take the
LowEnergyScanResult alone instead of also taking in the data as a
separate argument. We only did that because the data was stored
separately.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987909
Original-Revision: 9b3b47f13c3c8b41c4a00fe4a2d2a2688b9a807c
GitOrigin-RevId: 62d7fff8fd3938110701f4d685a7dd7cc4c83935
Change-Id: Iabf820d753bf3097e7d6261b40cc60054e615644