[roll] Roll fuchsia [audio][device_registry] FormatSets are static
To reflect the reality that RingBufferFormatSets (and
DaiFormatSets) are static -- they never change during the
lifetime of a running audio driver -- the Control method
`GetCurrentlyPermittedFormats` is removed. The field in the
static Device Info struct that contains this value (which
is renamed to `ring_buffer_format_sets`) contains the values
that were retrieved during driver initialization and will be
supported at all times while the driver instance is running.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

BUG: b/323273941

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999195
Original-Revision: cd5a6e48b694809bc9865ffcde42c7a1ed4e5861
GitOrigin-RevId: 696728933ac468b96df0b880d432ec74c314b20b
Change-Id: I11b823bfddd56406afc675e3e3ad9046c2a57ae1