[roll] Roll fuchsia [audio][device_registry] DeviceInfo fields
Add three fields to DeviceInfo, and rename another.

Rename `supported_formats` to `ring_buffer_format_sets`, to
clarify what type of formats this references. Note that this
is ONLY in the Info table now (not also retrieved by method)
since this is confirmed to be constant over the lifetime of
the driver instance.

Add `is_input` as a static field, even though this can currently
be deduced from DeviceType. This will not be possible for Codec
or Dai devices (in subsequent CLs).

Also add signalprocessing topologies and elements. These are
static properties that we can offer upfront in Info, rather
than requiring the user to explicitly retrieve them.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

BUG: b/323273941

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999196
Original-Revision: ca3e110bf2e4f29e05e631756efcecdc54d0c273
GitOrigin-RevId: 530ca62714f007b2282adef458405dfc0d84ee24
Change-Id: I0c71f42d8bba63bece0ae2e45328d9cbacd28ea6