[roll] Roll fuchsia [bt][hfp-hf] Translate +CIEVs to typed values.
The AG sends +CIEV indicators with an integer index and value.  The
mapping of these indices to specific indicators is originally specified
in a +CIND sent as part of SLCI.  This change adds a module to store
these indices and map them to typed AG indicator enums. It also adds a
ProcedureOutput variant to set the initial values of these indicators.

Glue code to use this module and ProcedureOutput variant is still to
come.

Test: fx test bt-hfp-hands-free-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/895561
Original-Revision: 89562b7a0c9a0aaa49c41b4224c5e99de99aacf0
GitOrigin-RevId: 9bd0c299d316a3c6719411640e4024f23a081a3d
Change-Id: If6924db238631d74f4eb6ce2ea0b7ae143812156