[roll] Roll fuchsia [bt][core] update LowEnergyAdvertisers to use EmbossCommandPacket return type
This change also fixes the
LowEnergyMultipleAdvertisingTest.AdvertisingHandlesExhausted test. This
test is meant to advertise up to the maximum number of advertisements
and then try once more to make sure the next one is rejected due to
exhausted advertising handles. However, in the case of
ExtendedLowEnergyAdvertiser, we use the default maximum number of
advertisements for its advertising handle map:
hci_spec::kMaxAdvertisingHandle + 1. In the test, we were going up to
only two so the third attempt would succeed. It was supposed to
fail. This change fixes the test to use MaxAdvertisements() instead so
as to ask the advertiser what its maximum number of advertisements is
instead of assuming it.

Original-Fixed: b/312896684
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984735
Original-Revision: d7e7d7e9560f0f4c2383823de12e79cfcdb8d901
GitOrigin-RevId: 15eb9cfd19f72669d63af49f917d73dae8e97748
Change-Id: Ifd7055aff86db78fe7af2524e46bfd9aa3760fe3