[roll] Roll fuchsia [fuchsia-async] Always attempt to read channels when polled
In b/303919602, there's a need to drain channels when shut down.  I
believe this is the only way that this will be possible.  Prior to this
change, it would be possible for a port packet indicating a channel is
readable to not have been processed, which would mean we'd miss a
message.  With this change, so long as all futures with channels are
polled, the fact that a port packet hasn't been processed yet won't
matter.

The same change is made to OnSignals: polling the OnSignals future will
always poll the kernel to see if the signals are set.

This might have a performance impact where futures get polled when they
haven't been woken (which can be the case with some future combinators
that don't track wake-up notifications separately), but that should be
rare.

Original-Bug: b/303919602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990573
Original-Revision: fba5c100f53927d4a080ce082d601a9797122a64
GitOrigin-RevId: 5c441edb1925934e74f1de0b17a6055785d22893
Change-Id: Icc50c34361e4401bed6a7787a5a87ded71a957d0