[roll] Roll fuchsia [fxtest] Don't hide the cursor
Previously, hiding the cursor was important because it would flash when
the status area updated. Now that the status updating code works better,
we do not need to hide cursor. Instead, fxtest "parks" the cursor in the
bottom right corner of the terminal.

This change makes the integration with zxdb easier because when zxdb
catches an error, the cursor is now visible at the zxdb prompt.

Original-Fixed: b/322420507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001901
Original-Revision: efcdf02bbe1530c44b20052e67fce2b89278180c
GitOrigin-RevId: 2975a3f69c052498b912bc61b66acfafac7d488a
Change-Id: I1fe6037ce1d9d358478550800de82e0aea401330