[roll] Roll fuchsia [idk] Remove dead rust code.
//build/sdk/meta contains a rust crate full of types that model the IDK
metadata. While these are nicer and more strongly typed than the
python code that manipulates the IDK, there's a problem: they're not
really used.

In fxrev.dev/998794, chok@ was under the reasonable (but incorrect),
impression that adding a new type of atom to this library was both
necessary and sufficient for adding a new type of thing to the IDK, but
unfortunately it's neither. Clearly, this code is doing more harm than
good.

Some of this library is used, specifically: code to do with product
bundles, and code that ffx uses to locate binaries.

This deletes everything else. RIP.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001513
Original-Revision: 39d76ddf3113c9f070dc44a469c09a3b9db237e5
GitOrigin-RevId: b5098fac05363e5c2e01e16279fdcffca4815788
Change-Id: I669e28abef42d7bbe85695d7754ea17bd6857085