[roll] Roll fuchsia [netstack3] Move InstantContext to base crate
In preparation to define the new timer context directly in the base
crate, move InstantContext there since it's a direct dependency.

Move everything that is itself a dependency of InstantContext.

Introduce a testutil module to netstack3_base so the fake
implementations of the instant contexts can move with the trait
definitions.

Original-Bug: 42083407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001655
Original-Revision: 9cd4fa387a84842aedb2faa82cd57223a0246e34
GitOrigin-RevId: 4faffa807ab8c58fe30babdf91c57fdfb7f088dc
Change-Id: I19bf8e59a8bdfc0d957dc8fec9b936fb1d05210b