[roll] Roll fuchsia [power][broker] Update documentation for ElementInfoProvider
ElementInfoProvider serves Status endpoints and Power Level information,
including levels and plaintext names, to clients that are mapped to a
specific Power Element via the Element's plaintext name. These names are
constants within the Elements and/or their managing components and
should remain constant between reboots of the same build, but may change
between boots of different builds due to changes in the codebase and
consistency is not guaranteed in that case. This CL updates the
documentation for the ElementInfoProvider tables to clarify this
information.

Changes:
- Update documentation for ElementStatusEndpoint, PowerLevelName, and
  ElementPowerLevelNames to clarify names/identifiers are unique between
  element and only guaranteed to be consistent between boots of the
  same build

Original-Bug: b/311470566
Original-Bug: b/281900115

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996169
Original-Revision: 0daeba7ffafa007edaa956adb30d95938e126a21
GitOrigin-RevId: e65873879eddbbf0f82aa9c553eb48b28236fa2b
Change-Id: I6ccfd2b7e9e16369ec3eadb2ba24e3889f5f3020