[roll] Roll fuchsia [audio][drivers] Make Codec unique_id an array
The CodecProperties.unique_id field is intended to be a 128-bit
identifier that is optionally provided by a Codec driver.
In FIDL, this field is currently declared as `string:16`.
The other audio driver types (Composite, StreeamConfig, Dai)
have this same construct, declared as array<uint8, 16>.

The Codec definition is unworkable because if any of the 16
byte values are > 127, then FIDL treats the value as the
first byte of a multi-byte character and subsequently
misinterprets the byte-length of the field. Furthermore,
if any of the 16 byte values are 0, client or server code
might fall into the trap of truncating this "string"
without capturing all 16 bytes.

This CL redefines CodecProperties.unique_id to match the
other driver types.

BUG: b/323273941

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002143
Original-Revision: 2a0e42f1549b06477bbe9325c1b99859f8904058
GitOrigin-RevId: 4d8dc0c9a77353bbf5c1567629fabe3e58c0d836
Change-Id: I40fbb629a1776bb146559cc89edeae5f7b5cdeea