[roll] Roll fuchsia [diagnostics][sampler] Increase test Reader timeout
Reader has a default timeout of 10 sec for a component to
supply Inspect data. For tests in emu, this is sometimes
exceeded and causes flakes.

This CL adds a for-test config to Sampler that tells the
Reader to have an infinite timeout.

As part of this, SamplerConfig is changed to a builder
pattern so that other code which uses SamplerConfig won't
have to change in the future every time we add a config option.

This CL also adds a check to the tests to make sure the
"single-counter" test component is publishing Inspect data
before Sampler tries to read it.

Original-Fixed: b/320697654

Multiply: sampler-integration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002323
Original-Revision: 4e10f0bbbf2f9187a3ab295dc6ca6a01984585c1
GitOrigin-RevId: 29a1970d6b4272ac5f314cde6f85ed2d838a426b
Change-Id: I6acb6da9389bf973a202798dda6b92810022ab9e