[roll] Roll fuchsia [fxfs] add overwrite extent value type
To implement fallocate, we need to introduce a new extent value type
which represents a preallocated disk region that will be written to
multiple times. These extents don't store checksums, instead relying
purely on the journal checksum records emitted as part of writing. For
now, only the format changes are added.

They store a block bitmap which tracks which blocks have been written to
at least once, because before blocks are written to they are sparse,
logically zeros.

An advisory flag in the object attributes is also added to indicate if
this file has any overwrite extents. This will help optimize open and
write, so that in the vastly more common case of having no overwrite
extents, we can skip expensive checks.

This also flattens the checksums into a Mode type which stores the write
mode the extent is in. Extents that used to have Checksums::None now
have ExtentMode::Raw.

Original-Bug: b/293943124
Test: fx test fxfs-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987914
Original-Revision: 0322de47568c66b755461e93d7719cf6db337e7f
GitOrigin-RevId: a8e1e1c17aa0399b51def1d31bc5974b06757480
Change-Id: Id45ff0f993fa3d55b85d22fde61b8e0f18d8481e