[roll] Roll fuchsia [component] Thin out ComponentProviderError
We already have ComponentInstanceError::InstanceNotFound, which
is the error returned when a weak pointer fails to upgrade. That
contains the moniker, making it more friendly than Source/Target not
found.

This also makes it more palatable to use ComponentProviderError as part
of the error returned by a bedrock Router.

Original-Bug: 324870669
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003172
Original-Revision: 7540c679eff70100efa26d6f670e315263cb45cd
GitOrigin-RevId: 4fbec8caf0879a045a745a04ed780eb5fef89130
Change-Id: I12c9d8fb035e9e9d4b15bad8d6a0737397463b96