[roll] Roll fuchsia [debugger] Parse nested Rust tuple accesses.
Rust supports the syntax "tuple.0" (or a nonzero number) to access the
corresponding value of the Tuple. Zxdb supported this as a special case
and not a float because in Rust, numbers can't start with a dot.

But Zxdb's parsing still allowed the following digits to be parsed as a
float. So if there were nested tuple accesses like "tuple.0.1" it was
parsed as <"tuple"> <dot> <float=0.1>.

This adds a new context-sensitive mode to the tokenizer for Rust where
seeing a dot without a leading digit actively inhibits treating the
following characters as a float.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003815
Original-Revision: a471907b40917cc1a5274530ca34b45ac1ce0a84
GitOrigin-RevId: a04ac0c976d6eedf2477e9e2b7be1a1756faa42b
Change-Id: Ic213133250d48bbc75ffa39626d551453f2d91ef