[roll] Roll fuchsia [netstack3] Support sending to broadcast addresses
Supports sending Ipv4 broadcast traffic to the Ethernet
all-ones broadcast address.

This includes both the limited broadcast IP address (255.255.255.255)
and the subnet broadcast address (e.g. for 192.168.0.0/24,
192.168.0.255).

As a result, this causes the out-of-stack DHCP client to be able to
successfully Rebind on netstack3 (as it depended on sending UDP
broadcast traffic in order to do so).

Original-Fixed: 42148833
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970713
Original-Revision: 3c10b40b1a135255b10f91f01d1740bdf7d67936
GitOrigin-RevId: 8f5f1366795c3fd425a0859b060d5060f1325398
Change-Id: Ib26c206917b1a978a403ca2350200a753033a16b