[roll] Roll fuchsia [netstack3] Update the bug # referenced by a TODO
This is the last TODO referring to TCP dual stack sockets. The nature
of this TODO is not critical to the functionality and should be a
cleanup. Update the reference so that we can mark the umbrella bug as
closed.

Original-Fixed: 42085913
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003259
Original-Revision: 50ee6a4a7b69b69d812e14a81e6026efdac4f6fa
GitOrigin-RevId: 111a91973c0d1752022aa729f6205c7e0b38ba9f
Change-Id: I3604d842cb3b7aac9fac2e497b147db06e955402