[roll] Roll fuchsia [phys] Ensure physboot-elf-kernel-test has uncompressed kernel package
As an interim measure, physload reuses the space previously occupied by
the kernel package (i.e., the STORAGE_KERNEL ZBI item) for the hand-off
to the kernel. With profile instrumentation, this hand-off further
encodes profiles which can take up a fair amount of space. When the
kernel package is compressed, it is too small to fit a hand-off with
profiles, so we reserve more space by leaving it uncompressed for now.

We also address a cosmetic image metadata bug that would incorrectly
present uncompressed ZBIs as compressed.

Original-Bug: 325513615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993241
Original-Revision: 3e30b9eb8e6182387d5cbcbcac091788358fe60d
GitOrigin-RevId: 448dc4d803db004787e2f9b03495c2e5cb6d889b
Change-Id: Iae91c3687bec81acb8027cf0c79bf926c44b3fa1