[roll] Roll fuchsia [amlogic_decoder] Fix bind tests and build rules
The build rules for the amlogic_decoder driver bind rules were organized
such that the tests were only defined in the host toolchain and only
when the target_cpu (within the host toolchain) was set to arm64. In
practice this meant the tests only ran on arm64 hosts (which are quite
rare) and as a result the bind test have bitrotted. This moves the build
rules for the bind tests out of the conditional, since they should pass
on any host architecture, and updates the test expectations to match the
current logic. The test now passes on arm64 and x64 hosts.

Original-Bug: 324437929
Original-Bug: 283475308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999673
Original-Revision: 1315949b9c658ffbd9ecd37607f6aa91af016149
GitOrigin-RevId: 1f70532cc7391e75a079d850570cee92a4381e8f
Change-Id: I588f781cb1f1497d566abe78b53f4c912b975755