[roll] Roll fuchsia [starnix/usercopy] Ret slice backed by dest slice
When `Usercopy::copyin{,_until_null_byte}` sees an address that is not
in the restricted address range, it terminates early and returns an
empty slice that it creates. The problem here is that callers of
the copyin methods assume that the returned initialized slice has the
same address as the uninitialized slice provided to the method.

This change fixes the copyin methods to always return an initialized
slice that has the same address as the provided uninitialized slice.

Original-Bug: b/328610022
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004134
Original-Revision: 3b09dd2e3035e733de6429d544fa37a3ed893746
GitOrigin-RevId: aaada13cea8e8fe8eca9467c99dcfb50b9fc2f0e
Change-Id: Ia6ed33b83d56132649feae1b08ee063e3126283f