[roll] Roll fuchsia [amlogic-display][viu] Do not pass canvas index as an argument of FlipOnVsync.
Previously the canvas_idx field of the display image was always
passed in to FlipOnVsync() to flip the active frame buffer used
by the VIU.

However, the canvas_idx field is actually invalid (and unused) when
AFBC is enabled. A better way to get the field is by directly
checking the `info` struct, and only when the image is not AFBC-
compressed.

This change fixes issue by removing the canvas index argument
from the FlipOnVsync() method.

Test: no-op.
Test: on VIM3 with AML028-30MB-A1, "display-tool squares" works.
      Note that "display-tool squares" generates non-AFBC frame
      buffers, so the modified code is tested.
Original-Bug: 323060335
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004234
Original-Revision: 1c420efae1b1cc2c8a0a4142c7abc38d8cab8a65
GitOrigin-RevId: e84c1ff4ec5f7fba5d504235e826696849c9ee3a
Change-Id: I286c2c7f3f6bc3c89fd9ac89a4caa16fe4db9198