[roll] Roll fuchsia [display][banjo] New struct for SetBufferCollectionConstraints() args.
DisplayController.SetBufferCollectionConstraints() currently reuses the
Image structure to convey the expected usage of the buffers in a sysmem
BufferCollection. This is misleading, because one of the fields (the
imported image handle) is not meaningful, and other fields (width,
height) are not necessary.

This CL reduces the confusion by adding a dedicated struct for
SetBufferCollectionConstraints(), and by updating the method comment to
clarify the implementation's responsibility.

Original-Bug: 316631158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003294
Original-Revision: 3dd86493ef3b8fd82df64a3118c5b88601c6633b
GitOrigin-RevId: fc46c7fa761fc6192207573d1191d4d1d169a55d
Change-Id: Ib3ad5bb0312a74818a6f958d9213890811b5646b