[roll] Roll fuchsia [botanist] Only send keepalives when no activity
Instead of sending a keepalive at every tick, only send it when we
observe no activity from the server since the last tick.

This is a cheap way to get to behavior that is closer to what openssh
does.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003752
Original-Revision: 3f3bc4f32935a3838fcef8d90adb3b141343a06b
GitOrigin-RevId: ce6180970525654ee67213ec010191f2213c6028
Change-Id: I2d990e2117ce881d5a67934408b611f1f60f6a23