[roll] Roll fuchsia [pkg-cache] Change how the dynamic index will be disabled
The dynamic index is still enabled, this just changes how the flag to
disable it works.

Previously, the dynamic index was disabled by ignoring it during GC.
This does not work, because the dynamic index verifies state
transitions during Gets that will start to fail when we use the open
package cache instead of the dynamic index to short-circuit Gets.

Now we disable the dynamic index by, when the flag is set, having all
mutating operations on it do nothing, so it will always be empty.

Original-Bug: 327208811
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004133
Original-Revision: c33e19363ac233f6be9d8cb9df460fc0e30551ad
GitOrigin-RevId: 1dc9878ba72dba4c94b0a76fc68fe70037e2c9f3
Change-Id: I044177223d9a9fd03fdd09ae60f703b87574a55d