[roll] Roll fuchsia [testrunner] Change "unable to record output files" to debug log level.
The log appears when trying to collect test outputs, which would not be
a reason for a test failure, so we should change the log level to be
less confusing. Also, this log will appear if there are no output files
to collect too, which is not an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004237
Original-Revision: 7794a3206a0bf0f231fb7bd497338d8be50752aa
GitOrigin-RevId: f0da69acdeabda76dbce34a99f7f9d1fd7a9ff6b
Change-Id: I94d3e58878b0540a61dc0b25bc75d69bf8aeb2b4