[roll] Roll fuchsia [tee] Add tee_internal_api header files and compilation tests
This adds the header files defining the TEE Internal Core API v1.2.1.
There are tests in tee_internal_api to make sure that the header files
compile in both C and C++. This also adds a Trusted Application that
returns success on every call.

There is not support yet for loading or executing these apps but the
build artifacts can be studied.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002112
Original-Revision: 681722daac132234323a650400bcc30c198829f9
GitOrigin-RevId: 684701308e392215e5275ff1e12a80af18eaa4be
Change-Id: Ida4105721cb5f03b2c79e5a555f00deb92155b68