[roll] Roll fuchsia [driver_host][assembly] Use vdso next when driver specifies
This is a temporary solution to enable drivers to use restrcited
syscalls. We will need a solution that doesn't affect all drivers in the driver
host as well as works for drivers that don't specify colocate = true in the
future.

Test: Manually added `use_next_vdso` to a driver an validated the the driver
Original-Fixed: b/328298860
host it was loaded into was `driver_host_next.cm`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000120
Original-Revision: 04e4e6bf3955b6b33a033f904e41450dbecb178f
GitOrigin-RevId: 6fd9a0a7e1daf9897d48a2ccc6408ca480b0a650
Change-Id: Ifcb1181b368d2f67f600dcf815c68b4a94bcfa13