[roll] Roll fuchsia [bedrock] Introduce libavailability
It is intended to be a minimal library used by libsandbox
when Router is moved there.

librouting also depends on it to serve legacy routing.

While doing this I discovered that OfferFromVoidToRequiredTarget is in
accurate, and is not used by bedrock. I didn't remove this out of
concern that legacy routing might be relying on this error to stop
routing early.

Original-Bug: 324870669
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004135
Original-Revision: 017f00311a3b62b9358ba21632677c9d480127f9
GitOrigin-RevId: 48b2005fa3482271c065a4a84be0eda4d4b6f791
Change-Id: I47214b7be2fb28d61dbfdee651818bf71b3bfc04