[roll] Roll fuchsia [ld] Make LoadModule use DecodedModule by composition, not inheritance
This is the next step in fully separating DecodedModule from
RemoteLoadModule.  It's now possible to instantiate LoadModule
using a pointer to a DecodedModule instead of embedding it.  This
is the mode RemoteLoadModule will ultimately use, though for now
it continues to embed DecodedModule.

Original-Bug: 326524302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002914
Original-Revision: 9e06a53d5392198ee5a5ebe5d9be9c54c2fbefc6
GitOrigin-RevId: 315c6269aaf149718dfdf757de389fc5e5a8ea18
Change-Id: I789e2fb72db35e9c859672fea04a9765ebaf5a67