[roll] Roll fuchsia [starnix] Fix Perfetto FD leak
Previously, we were using the sys_socket syscall to create a socket, but
that syscall also adds an FdNumber to the current task. We were not
closing that FdNumber. Also, the code that obtained the FileHandle from
the FdNumber was racy because some other thread could change that
FdNumber in the FdTable between the sys_socket call and this thread
grabbing the FileHandle

After this CL, we use the new_socket_file function to create the socket,
which returns the FileHandle. The FileHandle is reference-counted, which
means the socket will be cleaned up automatically once there are no
longer any references to it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004682
Original-Revision: 2e8150e5ec30e00c2c45a4d15117e4ad182a2ed3
GitOrigin-RevId: e04bbea1ce475f9758cc35ced08b3b3f6f21ed8e
Change-Id: I470daa3a64e4e21812c920c1f835e89f538697a9