[roll] Roll fuchsia [component] Properly wait for component start in tests
Currently these tests just happen to work because the async executor ran
things in the "right" order, but will fail as soon as that is changed.

The lifecycle tests need to wait for the component to be started before
calling `abort_controller`. Otherwise the MockRunner can't lookup the
controller KOID.

The destruction related tests might recreate a component with the same
URL. That causes `wait_for_url` to assume that the component is already
started so now the start functions call `reset_wait_for_url` to
forget the previous component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002576
Original-Revision: 662329486618fb632f5b88b6d0377d465f23193e
GitOrigin-RevId: 5f597c71cfa420e0c2f1850c1dba42da671cba54
Change-Id: I38adc7a15aea2272319e8d1d593926811e64f515