[roll] Roll fuchsia [pkg-cache] Use open packages to short-circuit resolves
This behavior will be enabled when dynamic package protection is
disabled (both of which are controlled by the protect_dynamic_packages
structured config flag).

During Get, pkg-cache does not check that all blobs are resident for
packages it knows must be cached, which saves reading the meta.far and
listing blobfs contents (repeated for each subpackage). This CL
replaces the dynamic index with the open package cache for Get
short-circuiting. This will slightly slow down resolves of packages
that would have been in the dynamic index but are not open.

This CL also adds the use of the cache_packages index to
short-circuiting.

Original-Bug: 327208811
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004550
Original-Revision: d6c30359225ce36499556c485adee047b00de6aa
GitOrigin-RevId: b6b35a8b8b5b7eee2932aa35b8937149fce35168
Change-Id: Ib0ed9e0f00ea8e9847714473de6368700a2520b3