[roll] Roll fuchsia [component] open_outgoing works when component is stopping
Previously the execution lock is held while we are stopping the
program. That will deadlock if anything during stopping the program
wants to use the execution lock, such as draining the namespace.

It appears that the execution lock was used to serialize
stop/start/shutdown. Now that we serialize them in the actions system,
there's no need to take a long execution lock.

This lets namespace/routing code acquire lock on execution during stop.

Next we can move some code that was taking the state lock to use the
execution lock and other simplifications.

Original-Fixed: 322564390
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004812
Original-Revision: 557ba4e193b164d03118a99604d05cebb0f25102
GitOrigin-RevId: f916892d6f49902669ac57bf0d3d072f06d21456
Change-Id: I505a1c2699ef762e2567c69e277115e95b3305bb