[roll] Roll fuchsia [debugger] Implement weak attaching to a process
This wires up the new information in the backend.

Weakly attaching means that the server will never send the process'
module list to the client until it is asked explicitly via `Modules`. If
a weak filter is installed, and matches a process at startup time, this
means we skip over the loader breakpoint and do not suspend the process.
Once the client requests the modules the first time, the attached
process is no longer considered weak.

When attaching to existing processes, an explicit `Attach` request is
sent, at which point the client may decide to strongly or weakly attach
in the Attach request message.

A strong attach is identical to the existing behavior and remains the default
behavior. The process will be immediately (or soon after being spawned)
suspended, the module list will be collected and sent to the client.

Several seemingly unrelated test cases had to be updated, which did not
explicitly insert a module into a spawned process and then injected an
exception for the session to handle for that process. The RemoteAPITest
fixture is actually injecting real ProcessImpls to use, and so we must
actually inject modules to the process. Before, receiving an exception
at all meant that we were attached and therefore symbols were loaded,
but that is no longer true. Most tests are migrated to use a new
function that creates a process for them and also a mock symbol module
so that they don't have to worry about making an additional function
call to install a fake module . More complex tests that configure their
own modules remain using the existing APIs.

Original-Bug: b/325281869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003656
Original-Revision: 534a61053c54c048640d5298d8337619a528e251
GitOrigin-RevId: 08e02c4dd4165c449c4142980f727123b02dd26b
Change-Id: I72e5d2f91e000dcd1a47ec60083959359ead52b6