[roll] Roll fuchsia [physboot] Provide page aligned memory ranges to zircon
On handoff we used to pass the memory ranges from the
booloader/boot-shim to zircon to preserve the reserved ranges.

In the case of devicetree boot shim, which use the memalloc::Pool
normalized ranges, this is no longer the case, which means the kernel is
not able to subtract the reserved ranges from the free ram ranges. This
is important, because by rounding down a free ram range's addr to a page
boundary, we might touch reserved memory.

As a result this change does three things:

(1) Memory Ranges of type peripheral are rounded down on the base
address and rounded up on the end of the range. No MMIO ranges are lost
in the process. Even with 'reserved' regions, this is harmless, since it
just means an uncached mapping of a part of page is created. (unaligned
peripherals).

(2) Memory ranges of type free ram's base address are rounded up
(trim unaligned head) and the range's end is rounded down (trim
unaligned tail). This protects possible reserved ranges from being added
back as free ram, when the original reserved ranges are no longer
available.

(3) As a result of (1) empty free ram ranges may be introduced, which
need to be elided (e.g. a fragmented bootloader allocation that is less
than a page such as the original devicetree).

Original-Bug: 300983188 # Fixed
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002856
Original-Revision: 1c58cd66fc62122466f3f673b3dfc8be7846089d
GitOrigin-RevId: e2f5ed9e1f64ede94bfc91d85ce72f8e8f6bafc8
Change-Id: Ie0035e27011c4e1071644bb31cc224bec495d567