[roll] Roll fuchsia [netstack2] Add logging for Terminal(User)
Add an overabundance of logging to help track down the cause of the
"Terminal(User)" flakes.  The goal is to log the key flow control paths
of interface creation and teardown to understand the order of operations
that's leading to the failure.  I expect this is far more than is
necessary, but I don't want to have to go through more rounds of this if
I can avoid it.

The logging should also be inexpensive enough to not cause issues.
These RPCs are administrative, so there aren't tight performance
bounds, and we're only adding a handful of syscalls.  We're also not
adding any new failure points.

Original-Bug: 328297563
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005852
Original-Revision: d0cc2fae660f41d9386ce7d05675a1f626815f8d
GitOrigin-RevId: f57b1e99e15d1e4b832fdfc53f487869f66eac0d
Change-Id: If82b79adf0bf99c77ef2d8a8e5452e778077efab