[roll] Roll infra/recipes [api_level_roller] Use new script
We moved functionality for this roll into a separate script in the
integration repository, which itself handles interaction with
version_history.json and CIPD to calculate the correct set of manifest
changes.

Since we don't know if that script performed any changes, we will
unconditionally attempt a roll and rely on the auto_roller module to
skip the roll if there are no changes.

Original-Bug: 319136115
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1001925
Original-Revision: aa4e3b565e3a237711f1cf4a56d7165218507903
GitOrigin-RevId: 3c527ed626f9f2c476a1909f1feeead199fccc74
Change-Id: I43b9959736fb494d7cebe6d4f3a4fb44c99ceec6