[roll] Roll fuchsia [bt][obex] Assign a ConnectionId if requested
The remote OBEX client may request a directed OBEX session to a specific
Service by including the Target Header with UUID in the CONNECT request.

Assign a Connection ID to the connection if requested. This ID will be
included in all subsequent OBEX operations and is fixed for the lifetime
of the OBEX connection. Because there is only one OBEX server for an
underlying RFCOMM/L2CAP transport, a static ID of 1 is assigned. This is
mainly a formality and is irrelevant in almost all situations.

Original-Bug: b/42080293
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005372
Original-Revision: fceb20510cf5a2ea8e1ca2f59ce3e701061cc4b8
GitOrigin-RevId: a591bbe0583eaaab598bd3b20b78ba377e88356b
Change-Id: Id0e6472ab87c892ec56fb2241f1fa709f0ae3344