[roll] Roll fuchsia [fdf][devfs] Add DevAddArgs Controller Override
Provides a temporary way for Node to allow the
Compat shim to serve the fuchsia.device/Controller
interface for shimmed DFv1 drivers.  A new field is added to
the DevfsAddArgs which provides a connector to a Controller interface.
If provided, that controller interface will be used instead of
the Node's implementation.

bug: 320726737
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005017
Original-Revision: 346ab601100bb3e886a08051bea643202a29d505
GitOrigin-RevId: 94a279d24bf8603c7a9172195aa83b1306d7f693
Change-Id: I0239c238a974f812dfa235355b21fab498dcae2d