[roll] Roll fuchsia [netstack3] Fix socket inspect local addr any
Print the address correctly for TCP sockets that are locally bound to
the ANY address.

As `SocketAddr`'s type definition forces an address to be specified,
it was necessary to generalize to allow the unspecified ANY address.
Added `AddrAndPortFormatter` for formatting an address and a port, and
reuse this implementation in `SocketAddr`. Modified `AddrAndZone` and
`ZonedAddr` where necessary in support of this.

Modified the existing socket integration test to assert this.

Original-Fixed: 326283980
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002858
Original-Revision: 979183505353074bd6ad66cfa1b39e0fcbbf97a6
GitOrigin-RevId: eebd1d3b8013b1009a6b26e5b0685b136eb98ccb
Change-Id: I10b79c5858f922f0af80fee648cf097e0c46ff4b