[roll] Roll fuchsia [ffx][writer] Adding VerifiedMachineWriter
This introduces VerifiedMachineWriter as a struct
that effectively extends MachineWriter to allow
the production of a JSON schema describing the machine
output.

A partial implementation of a bespoke schema is partially
replaced in this CL due to name collision with VerifiedMachineWriter.
The rest of that implementation will be removed in a follow-up CL.

Original-Bug: b/326125002
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006493
Original-Revision: f89590c0ef48ecbc6457bb873833e930b772bd9f
GitOrigin-RevId: 2a74f9961cc4402facd17050dcc0c4f839783f74
Change-Id: I8475eb06ee961dd82cda04d9639acaacbca12804