[roll] Roll fuchsia [ld] Split out <lib/ld/remote-decoded-module.h>
The <lib/ld/remote-load-module.h> header is getting large.  Split
out <lib/ld/remote-decoded-module.h> for the RemoteDecodedModule
class, which is getting to be somewhat more independent.

This does nothing but split part of the contents of the existing
header verbatim in the new header and adjust the `#include`s to
the subset each header needs.

Original-Bug: 326524302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004912
Original-Revision: 63ef050426d741863361332a0dd5fd12bbbc90c8
GitOrigin-RevId: 8996574d3055e492ab1dabbfde5eb3c83aa36aea
Change-Id: I9685544cb75bf0d2db47530fca4e0391e1ed4311