[roll] Roll fuchsia [fuchsia-fs] Add another way to read files
The fastest way to read the contents of a file is to:
 - open the file with DESCRIBE
 - get the stream from the OnOpen event
 - seek to the end of the stream to determine the file size
 - read that many bytes from the stream

The problem with the existing fuchsia_fs::file::read method is that the
file has already been opened and now Describe needs to be explicitly
called to get the stream. Most clients get the FileProxy by calling
fuchsia_fs::directory::open_file and then only use the proxy to read the
contents of the file. Combining open and read into one method gives the
best performance.

Original-Bug: b/324239375
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999712
Original-Revision: 5a44748720405794aef246821c9e877f062996d2
GitOrigin-RevId: 1b5c72dae532b8e35bd614189c793fcdfbcf0d7e
Change-Id: I663dbc0840449623e5ede717e1372b9b37c58b1d