[roll] Roll fuchsia [display][banjo] Merge clamprgb protocol to display controller protocol.
The fuchsia.hardware.display.clamprgb library had an optional
ClampRgbImpl protocol for display engine drivers to set the minimum
RGB value to be shown on the panel.

This was designed to address a hardware issue on a specific board
so the protocol implementation was optional.

Having multiple banjo APIs makes it difficult to migrate the driver
to DFv2. To make it easier for the DFv2 transition, this change merges
the ClampRgb protocol to the display Controller banjo protocol;
SetMinimumRgb() now must be implemented for all display engine drivers.

This doesn't change the FIDL fuchsia.hardware.display.engine API. The
design of the ClampRgb method in the FIDL API will be addressed in a
separate change.

Test: no-op.
Test: On Astro, "display-test --clamp-rgb 128"
Test: internal product on Astro works.
Original-Bug: 328903017
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004583
Original-Revision: dfc21fa736a7df2342c2ddbd689457c072f41268
GitOrigin-RevId: b3fa9a288e21d43959e510e21c61668652f7079e
Change-Id: I3e9057eff6b4e4205cc5a8d684d5bec16003792b