[roll] Roll fuchsia [display][banjo] New struct for ImportImage() args.
DisplayController.ImportImage() currently reuses the Image structure to
convey the expected usage of the buffers in a sysmem BufferCollection.
This is misleading, because one of the fields (the imported image handle) is
not meaningful.

This CL reduces the confusion by adding a dedicated struct for
ImportImage(), and by updating the method comment to clarify our
expectation that this information will be sourced from sysmem at some
point in the future.

Original-Bug: 316631158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004841
Original-Revision: 08c6c6cb80ca9b980f9848dab0f950174765a03c
GitOrigin-RevId: d7a2e1ab4392ea3352c87bc3de49b86d72c3bc13
Change-Id: I31cf88c12adcf1529492ede08ae3c4b7ee940259