[roll] Roll fuchsia [logs] Add ordered mutex
For both the old and new symbolizer we want to
ensure that futures are polled in a particular
order at a synchronization point, and data
at that point is guarded. This is similar
to a futures::lock::Mutex but is different
in that it is optimized for single-threaded
contexts and guarantees tasks are unblocked in the
order that they called lock.

MULTIPLY: host_x64/ffx_daemon_log_lib_test

Original-Bug: b/299980894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006876
Original-Revision: 02f07142681d4e529ec3c8cb57047cd15bd8013b
GitOrigin-RevId: aa282503970c9e2990970d417e596152876db5a0
Change-Id: If4f076762b93e661c1432fa3910a1fe3e37b9ddc