[roll] Roll fuchsia [bt][obex] Allow duplicate headers in multi-step requests
Update the HeaderSet to allow for the adding of duplicate headers to the
set. A duplicate header is a Header that exists in the collection with
the same header value. If the value differs, then adding the header to
the collection is still an Error.

Original-Bug: b/42076096
Test: fx test bt-obex-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006716
Original-Revision: fb2e70471a1067ae0f1234f1197dc173e9341bee
GitOrigin-RevId: 3b7558e63e6c68605340fac4bdc3b4c4a9cf4f98
Change-Id: I45f97a09deee90ecca2fc0b25ab4aef19cf5316e