[roll] Roll fuchsia [fxtest][debug] Wait for debugger to be ready
zxdb has a small startup time to connect to the backend on the target
and install any filters for test components that will be spawned.
There's a race condition between TestManager spawning all of the test
cases and DebugAgent such that a test that fails sufficiently fast may
not be caught by the debugger, or worse, crash the debugger due to some
strange state that would otherwise be impossible.

This makes use of the existing support for zxdb's signal-when-ready
argument.

Original-Bug: b/325281869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006057
Original-Revision: c961306744f02d9c038ab27e541d4dff06e231cd
GitOrigin-RevId: 285877fb2c9b59ba2b7ab039ddb4a2a6ce80750f
Change-Id: I277af95537f7cbe5896eefe5f3af9a990b370211