[roll] Roll fuchsia [zircon] zx_sampler_*
This CL Proposes/Creates a @next, build argument gated, syscall enabling
sampling data from user threads.

Currently, a user program can sample a given thread by obtaining a
handle, suspending the thread, then using process read memory to iterate
over the targets stack. This process takes ~2-4ms on supported hardware
and is too slow to be useful.

This CL introduces an approach for the kernel to provide assistance for
taking samples. Creating a sampler causes the kernel to allocate and
return a buffer to the user that it will write samples into. The caller
can then request a thread start being sampled. This causes the kernel to
(when safe), sample the thread's user stack and write it to the buffer.
With this kernel assistance, sampling overhead is now ~2-4us per sample.

These syscalls are intended as a temporary solution to unblock
performant profiling on Fuchsia as well as a building block and
iteration opportunity to design and build a robust set of observability
apis for Fuchsia.

Original-Bug: b/301632554
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939358
Original-Revision: 37ff62c616bf4f3d5f433379335fa9dc8baa7516
GitOrigin-RevId: 2d227ca8acbed725438c2e48f578eedf43eae616
Change-Id: Idd051ace33180077a7e26907adee79c734606806