[roll] Roll fuchsia [sysmem2] Add R8G8B8X8 and B8G8R8X8 formats.
Adding R8G8B8X8 and B8G8R8X8 to make X vs. A explicit, instead of
relying on participants to communicate X vs. A separately.

In some cases this will end up mapping to "A" in vulkan since vulkan
doesn't have X vs. A at format level, but at least this way we can
complain to the log (or similar) if the format says X but a client is
trying to treat as A.

BUG: 327498161
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005645
Original-Revision: 8cd7d141dc04c44f4f72cbc849496daadeac4d3c
GitOrigin-RevId: 28e5f99217878981badac767f48c4540035b58cb
Change-Id: I9875d4222ab42a875a393e737af00db2cf933ea7