[roll] Roll fuchsia [ffx][sdk] Adding machine writer to `sdk version`
This adds the use of VerifiedMachineWriter to the sdk version
subcommand.

This is part of a project to make all commands that are integrated
into other programs/scripts use the machine interface and provide
a schema of the output to enable identifying breaking changes.

Original-Bug: b/324937400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005637
Original-Revision: 42959ff6aba36748cd22c6665a0408ada4612958
GitOrigin-RevId: b8d6af8688675b95d633ea272499d2626f935bab
Change-Id: I9947411996f47ceaeb40e672f1e2d7ad151a855c