[roll] Roll fuchsia [kernel][usercopy] Support capturing access faults
Modifies the usercopy code paths to optionally capture access faults on
architectures that support explicit access faults (currently, this is
only ARM64). Note that this code path is not actually used anywhere yet,
so current behavior and performance should be completely unaffected;
future CLs will actually expose this functionality to the `user_ptr`
layer of the system.

This change is in service of improving the handling of
FutexContext::Wait in the thread lock breakup.

Original-Bug: 328674858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004052
Original-Revision: 40a9161076366eb13ebeda8951c9f15e4f4498f0
GitOrigin-RevId: d7794f74a0c77de1d91d393a51d2ac2786b7fdaf
Change-Id: Idaf06d6ec6ccf1df89160f768c453274d234f951