[roll] Roll fuchsia [graphics] Use fdio_service_connect to connect to Provider
We need to stop sending OnOpen events for service connections as part of
the io2 migration, as this is no longer supported in io2. This is
because it requires clients switch protocols over the service channel by
first listening for a fuchsia.io/OnOpen event, then switching protocols.

The VFS libraries now ensure an epitaph is sent on failed service
connections, which clients can listen for or handle as if the service
was closed.

Original-Bug: b/327261966
Test: fx test get_hardware_display_coordinator_with_provider_service_test get_hardware_display_coordinator_without_provider_service_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003772
Original-Revision: 310a299db7a40f16b54fbc5f4cda9fb26f84a102
GitOrigin-RevId: 89b8aac89b90b5c46cbd71359e300c82529b2f59
Change-Id: If1982a6fdb2d3fa4bc77bc36bcfae9ca51070dca