[roll] Roll fuchsia [ffx] Fix discoveryless partial-matching
When doing local discovery, there was a path for matching against
a specific name. But when using daemon-based discovery, partial
matching is supported. This CL implements more general matching
in local discovery.

There was a kink in the implementation: when specifying a string,
we still want to return immediately if we find that target name,
but if it's just a substring, we still want to capture all the
matching targets. This required re-working the stream logic
significantly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003954
Original-Revision: 5ad0b6f80de82bb73f0772260e84d07d48e5b9ad
GitOrigin-RevId: ca9e57e814601cf1a1eac6b35fb64866d5fec8b7
Change-Id: Idc41079097b920ba53ba12e2f7c2c03e26ca70b5