[roll] Roll fuchsia [rfc] Asynchronous operations are on objects
This proposes changing the semantics of asynchronous operations in the
kernel to be defined as operating on objects as opposed to the handles
referring to those objects. This modifies the behavior of the
`object_wait_async` syscall when the handle is closed and proposes a
new syscall for cancelling all asynchronous operations registered with a
given key.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985309
Original-Revision: ce26e3d979d0976d28cdb3368ea433570c62b3ec
GitOrigin-RevId: 8e6b09f10540f09c3dd6d0ce981428fd0b00d7c8
Change-Id: Iec07582ff445ec6592be6a8b23e44d7fe7ab3773