[roll] Roll fuchsia [devicetree] Add support for composite nodes without pdev
Sometimes we need to create composite nodes out of other nodes without
the need for a platform device. For example, pwm-init is one such node
which is a composite of pwm and gpio nodes, and does not need a platform
device.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006419
Original-Revision: ab551ba202103d8c1f9c6656ee71ac0e43caeba8
GitOrigin-RevId: cf9994c987216193212a6f0b50ed4c672a7857f0
Change-Id: Ifa0d4c69838331ba2cd00a1630a894c5f710b258