[roll] Roll fuchsia [forensics] Generate a unified binary & rename it
This patch reorganizes the forensics build so that it generates a unified binary, and uses renamed_binary template to generate a list of renamed_binary targets. The problem was originally discovered in the referred bug, but resurfaced in the LTO context, where we started
seeing a size increase because of duplication in forensics binaries.

Original-Bug: 323231755
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007796
Original-Revision: 1bb0a37e4a70c22372e91b250e93d51e9e9e1fbd
GitOrigin-RevId: 1c29d5aad6840f6b3e392e0007d8140934c644d6
Change-Id: I2af3de381dacd4922cd8a91d8205594ce34d7ff1