[roll] Roll fuchsia [bazel][docs] add tool to extract bazel docs
We cannot use stardoc directly because it uses rules_java which our
build does not support. Instead, we use the new rule that is introduced
in bazel 7.0 which extracts docs into a protobuf. This is what stardoc
is doing under the hood.

This CL has a stubbed out tool which will parse the protobuf and render
markdown to be consumed by our docs site. The implementation of the tool
will come in a later CL.

To extract the testdata proto run the following command:
`fx bazel build
//tools/bazel-docgen/testdata:extract_foo_docs.textproto`

BUG: b/306761829
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004838
Original-Revision: e62ada7e70a31a7d432efc1e4f2b943d53e41f5b
GitOrigin-RevId: 78837fa6c9287f641e8a46053535f5622bcd8355
Change-Id: I978bc22d8621661ad1f4f557ea4f827fb2c22be9