[roll] Roll fuchsia [superproject] Roll scudo [scudo] Refactor allocator config to support optional flags (#81805)
Instead of explicitly disabling a feature by declaring the variable and
set it to false, this change supports the optional flags. I.e., you can
skip certain flags if you are not using it.

This optional feature supports both forms,
  1. Value: A parameter for a feature. E.g., EnableRandomOffset
  2. Type: A C++ type implementing a feature. E.g., ConditionVariableT

On the other hand, to access the flags will be through one of the
wrappers, BaseConfig/PrimaryConfig/SecondaryConfig/CacheConfig
(CacheConfig is embedded in SecondaryConfig). These wrappers have the
getters to access the value and the type. When adding a new feature, we
need to add it to `allocator_config.def` and mark the new variable with
either *_REQUIRED_* or *_OPTIONAL_* macro so that the accessor will be
generated properly.

In addition, also remove the need of `UseConditionVariable` to flip
on/off of condition variable. Now we only need to define the type of
condition variable.

GitOrigin-RevId: e9d7da6c4cd0938c2f4b2b95bbec0b870fc7d08b
Original-Revision: 94ef308403a0e86215b6639604a70b4d1dd17aae
Roller-URL: https://ci.chromium.org/b/8753510386732391873
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006883
Original-Revision: 2e6044833bf6499eafc1d64ac8ce760fb559c330
Change-Id: I29c8376a3d62e6557e1ac455d6f2984793498ed5