[roll] Roll fuchsia [zircon] Fix providing ZX_HANDLE_INVALID to process_start
...when the bad handle policy is enabled. This system call is defined to
allow providing the value ZX_HANDLE_INVALID if no handle transfer is
required but the handle lookups used to implement this would trigger the
bad handle policy if enabled. This guards the handle manipulations so
that in the ZX_HANDLE_INVALID case no policy errors are raised.

This codepath is tested by ProcessTest.ProcessStartNoHandle in the core
tests and the Zircon Rust library wrapper tests but we don't currently
run either test suite with the bad handle policy enabled so these tests
do not report an error. We should run these suites (or subsets of them)
with the bad handle policy set to a higher enforcement level.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007714
Original-Revision: 67f68c23bb46e43fd09654561582c64e2baff792
GitOrigin-RevId: f966afaef01d5895ea4e1574c677424890c072e0
Change-Id: I45fa3facdd55a777246d869a36aff691a0fb2144