[roll] Roll fuchsia [network/bench] Wait for interface control handle closure
When cleaning up after running the "interfaces" workload in the resource
usage benchmarks, wait for the control handle to the interface to be
closed with a terminal event rather than just dropping the control
handle and waiting to observe removal on the interfaces/State protocol.
Netstack3 does not wait for all interface cleanup work to be complete
before emitting the interface removal event to State watchers, but it
does not emit terminal events on the control handle until all cleanup is
complete.

Synchronizing only on the State removal event meant that the benchmark
observed that resource usage had not subsided to baseline for Netstack3
even after the benchmark had finished running.

Original-Fixed: 322860526
Test: resource-usage-benchmark-netstack3
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007294
Original-Revision: 0666015e2304baac8d82faf219bdc65a44a0bd7e
GitOrigin-RevId: 53cd4ebe67d47a8677e8c809973add24296dce29
Change-Id: Ie5fa49a92f2a12587c7713b17e04324d5a9b0a6a