[roll] Roll fuchsia [lib/network] Fix race in or_terminal_event
There are many more details in the attached bug, but TL;DR: We're
pretending that the query and terminal event futures are sequenced, but
they're not. It's possible for both of the futures to become ready in
between the time the select() future polls the query future and terminal
event future. In this case, we return the terminal event, even though we
really want to return the query result.

The fix is to now_or_never() the query future when we hit the terminal
event case.  Since FIDL messages are always processed in order, if there
was a query response, it'll be ready immediately.

I ran the flaking tests against this >93k times and didn't observe any failures.

I'm punting writing a test to a follow-up CL because it's tricky to
write and this flake is occurring often enough that it's contributing to
tree closures.

Original-Bug: 328297563
Multiply: netstack-interfaces-admin-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007494
Original-Revision: 29dbe28ce0ee149e3cacd5691035f6445f9c8980
GitOrigin-RevId: ba6cd629da8fc40406d13d6e89b1f3bad5b4b0b0
Change-Id: Iaaed96797e50f2a8b151ff1cd8e41a05900c1022