[roll] Roll fuchsia [bedrock] Unify RouteFn and Routable
Routable is a trait that lets you route capabilities, so we can
implement Routable for all RouteFns. This matches the idea that Router
is the object-capability representation of Routables.

Also reorders decls a bit so that the most important types float to the
top.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007794
Original-Revision: cf651d64091166e50514bc4d2e597ae76c9ffc21
GitOrigin-RevId: 0588b18236feddc78c21a5e57a2813977df2073c
Change-Id: I2040282e9b1fdf83d2e78253e4bfc05f1a43bd92