[roll] Roll fuchsia [fidlc] Improve recursive compilation logic
This improves how fidlc recursively compiles decls that types refers to.
Previously, Typespace::CreateIdentifierType would call CompileDecl if
status == kNotCompiled && kind != kProtocol. Now, it always does unless
there is an `optional` constraint or `box` type to break the cycle.

The old behavior was strange because a given piece of logic should
either rely on something being compiled or not. If it relies on it, then
recursively compiling only when it's kNotCompiled is not good enough: it
should be unconditional (and hit ErrIncludeCycle if already kCompiling).

This change is in preparation for removing the SortStep. With proper
ComipleDecl recursion, the CompileStep already detects all cycles and
can easily build declaration_order, making the SortStep redundant.

This also removes some superfluous cycle detection logic in
Typespace::CreateHandleType and Typespace::CreateAliasType.

Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008212
Original-Revision: ce6d8c3712f2724e2eee9fb370b73978b4ea0ee4
GitOrigin-RevId: b3b3fd170bf724ba5e44cffab541117ca1d6b760
Change-Id: I558951ed8e02dc4dcd6d34cbb7a2a2e60fc5b832