[roll] Roll fuchsia [sysmem] Permit getting v2 Allocator from v1
This allows connecting to sysmem2 Allocator given only a sysmem(1)
Allocator. This helps in situations where shared lib code is being
updated to sysmem2, but gets created by passing in a sysmem(1)
Allocator, and the shared lib code doesn't necessarily know that it's
ok to connect to sysmem2 instead via /svc, such as when the lib code
is being used in a driver.

BUG: 306258175
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963656
Original-Revision: 1b1334cf801c7c281718f259dc7960fb4db505f9
GitOrigin-RevId: 208389a50b19eceba16a305bb8e5d1bfae6b05b0
Change-Id: I48b6b7a3f143a5198b9f2ab2c9711c9212d214f6