[roll] Roll fuchsia [botanist] Close ssh client if not using it for capturing syslog.
We create a new client for adding the package repository and then
capturing syslogs. However, now that we're using ffx log to capture
syslogs, that client doesn't do anything for the rest of the task unless
the target reboots and we need to readd the package repo. However, it'll
continue to keep sending keepalives to keep the connection alive. We
should just close it from the start so we don't keep sending keepalives
and adding unnecessary logging for a connection that mostly doesn't do
anything.

Original-Bug: 324285824
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007585
Original-Revision: a42ef933078875e23739d5a1f58dba0cd50bfb31
GitOrigin-RevId: 2ac56a54f03a0f3ce93263dd60552c2aeb61bb96
Change-Id: I8900b6132cec8f6b86d642968ac1db89b0a4b8fd