[roll] Roll fuchsia [kernel][vm] Prefer size_locked as virtual method
Users of VMOs sometimes need to know the size while they hold the lock
to prevent races. This change means that the lock is needlessly grabbed
when querying the size of physical VMOs, but these are a very small
amount of VMOs so this overhead not very relevant.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007912
Original-Revision: 9d6996ac963890b790c77c91c754e674bd39ad7e
GitOrigin-RevId: e42bbec7eeb4447e6c35e189d1d9d47be353537e
Change-Id: I2851752b16cf5e37872eedaab82bc701d35a408e