[roll] Roll fuchsia [lib/network] Add test for or_terminal_event race
The race was described and fixed in https://fxbug.dev/328297563.

We pulled the Control::or_terminal_event() method into it's own
freestanding function to make it easier to test without having to drag
in all of the FIDL client code.  It's also unclear if we can reliably
trigger this bug if we tried to reproduce it at the FIDL level.

Ran against the previous code to make sure it fails reliably when the
bug is present.

Original-Fixed: 329679745
Original-Bug: 328297563
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006854
Original-Revision: b3defea087a1303276dbe9b1a0218b774504edf9
GitOrigin-RevId: bbc2aa6f4013601692fa126f3cb7495fe6be9928
Change-Id: If50c949c3016bb9b59787bf206a29ed7b7cc5145