[roll] Roll fuchsia [fxfs] Refactor versioning scheme
The versioning scheme has been considerable source of confusion for Fxfs
developers.  This CL attempts to rectify that.

The common interpretation of FooV20 was that:
- Foo was changed at V20, and
- FooV20 represents the struct at that version.

This was inconsistent with the intended interpretation, which was that
FooV20 was the *last* version at which Foo looked like that struct, and
Foo was changed at some unknown prior version.

Based on discussions, this CL moves to the common interpretation; now
FooV20 means that Foo was changed at V20 to look like that struct.

I noticed while making this change that at least one version was done
incorrectly; new fields were added to ObjectKeyData in V33, but some of
the containing structs were not versioned.  This was corrected.

Original-Fixed: 326988259
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007514
Original-Revision: 665b589c4780124980f46a7755892a5a3645297c
GitOrigin-RevId: 9a71c8b9bf63dc328be379c1f738fe6211503a51
Change-Id: I6ff58128c257bc39e2b644d15d07f4bd38708baa