[roll] Roll fuchsia [bt][core] low energy advertiser general cleanup
This change includes a variety of cleanups to LowEnergyAdvertiser and
its related files. These cleanups are mostly in preparation for the
inclusion of extended advertising PDUs. The extended advertising PDUs CL
was getting quite large and having these cleanups in there was making it
a bit noisy and hard to focus on what exactly is new specific to just
extended advertising PDUs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008935
Original-Revision: 97b7f93b97135690dba0de3060d9c6ff3b34fce7
GitOrigin-RevId: 7b4de406e1fb15c25670af6f3494d07918607671
Change-Id: Ia618789c5eb5fb8af72d5edb0ea54aa8d55b3341