[roll] Roll fuchsia [netstack3] Local ingress hook
Add the local ingress entry point for the filtering crate, which occurs
for incoming IP packets that have been routed for local delivery but
have not yet been dispatched to the transport layer.

This CL modifies the existing tests for the ingress hook to exercise the
local ingress hook as well, since, for the socket layer, there is no
observable difference between traffic being dropped at ingress or at
local ingress.

It also adds a separate flavor of integration test to exercise the
difference: in this test, instead of a client and server directly
communicating, we introduce a router in between them, and install
filters on the router. In this scenario, installing filters on local
ingress does nothing, because all client-server traffic is forwarded and
therefore skips the local ingress hook; installing filters on ingress,
however, does affect client-server connectivity.

Original-Fixed: 326082742
Test: netstack-filter-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005312
Original-Revision: 8e356a08ff09e701b94ef0d95c7d19e21538967a
GitOrigin-RevId: 19772c54f3e5bee5b5e7e6c828e15314a1a73c46
Change-Id: I1c6db347d4a63ed2c6382eb2083cb0483da3d432