[roll] Roll fuchsia [bazel][boringssl] generate bazel rules during roll
Updates the roll_boringssl.go script to also generate the build files
for bazel. These files are used in the boringssl bazel repository. We
were generating these on demand but the python script spawns a process
which calls out to the go binary on the user's path which is not
hermetic in our build system.

BUG: b/327257847
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008842
Original-Revision: fa31ab89edf12e6e1d6621ffabafd3812eb75016
GitOrigin-RevId: df0e1aed3876fa2186581db38d131c0180e1f0c2
Change-Id: Ia80796409e19401b5c0ab66c02502c83c05be792