[roll] Roll fuchsia [netstack3] Remove OnceCell from netstack3_filter
Currently netstack3_filter stores a OnceCell on each uninstalled routine
to cache the converted version of the routine -- the version with its
"debug info" stripped. Instead of using a OnceCell, which has
unspecified behavior on reentrant initialization, eliminate locking and
use a regular HashMap to incrementally build an index of converted
routines. This has the benefit of allowing us to define the behavior
when encountering a routine cycle; currently we define this as a panic
because Bindings checks for this, but we could also choose to return an
error to the caller.

Original-Fixed: 325492760
Test: netstack3-filter-test
      netstack-filter-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007957
Original-Revision: 5c7d33dd10cdbb989bc132ebafad4e9d7910c691
GitOrigin-RevId: 4e9e0bca1f7ba203a1552e2d20df609e9c14d5cc
Change-Id: Ic729dd71b6e9b7cf05f023955df7aba180965bde