[roll] Roll fuchsia [wlan][policy] remove legacy file clean up
This change removes the code that was in place to delete the legacy file
being used for saved networks persistent storage. Before this, the
legacy file already wasn't being used if it had anything in it, just
deleted in case it was still around.

Original-Bug: 42166225
Test: existing unit tests for loading saved networks verify that the rest of the behavior remains unchanged. This change just removes file cleanup that we don't need anymore.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1008216
Original-Revision: 5ec7e55983697923aaeb3744c679ecd9a3c909df
GitOrigin-RevId: 2538f54ab6ba4aeca01a5742eb7abfba00e78a68
Change-Id: I7e6cfc1961b131b9a9eeef1dbc0eb2bf809d81f6