[roll] Roll fuchsia [sdk] Adjust documentation for SAG's FIDL
Updates the documentation to indicate _how_ these protocols are expected
to be used.

Additionally, we remove the passive deps on WakeHandling and
ApplicationActivity because we do not see a use for them today.

Additionally, we convert the level constants to enums to ensure that the
level values do not overlap and to indicate that the values are related
to each other.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006813
Original-Revision: 9575ab6cac9c0be31e9c90247409355cf02fb5a5
GitOrigin-RevId: 6f082677cb63d35554f13ae063740fc5b2e3257a
Change-Id: I9ca715aef915780d79193517a7a9ca27e952c4c9