[roll] Roll fuchsia [network/test] Shard filter FIDL test
Use the automatic test sharded to break the filter FIDL test up into 2
components so that each one is less likely to time out; the test already
has a long duration and there is a lot more coverage that will be added
to it.

Also, use a shorter timeout when waiting for an event we expect not to
occur. We're currently using
netstack_testing_common::ASYNC_EVENT_NEGATIVE_CHECK_TIMEOUT, which is 5
seconds, and in the filtering tests we have *lots* of negative checks as
we add filtering and verify that packets get dropped (and therefore,
e.g., a socket connection fails). Reduce the timeout to 1 second. This
has the risk of increasing the chance of a "false pass", where the event
was delayed by up to a second and therefore the test passed when it
shouldn't have, but it will still fail often enough to catch issues.

Original-Fixed: 330243476
Multiply: netstack-filter-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009375
Original-Revision: 3e513d4cbd1b97750d4b92012ae56078084be57e
GitOrigin-RevId: 0108b5756a40e0cb88fe59676ef546884ecf6903
Change-Id: I548919d1ff064cac175d904476615be2ef082909