[roll] Roll fuchsia [component] Increase MAX_NAME_LENGTH from 100 to 255
This matches the limits placed on a name to what's set in fuchsia.io and
what's set by the process framework on namespace names. It helps to
avoid arbitrary divergence between CF and the layers beneath, as we
change Dictionary key to Name.

Because UrlScheme validation references MAX_NAME_LENGTH, that's also
increased.

If a downstream customer creates a component with 101 bytes in a
capability name, that will compile with the `cmc` from the latest SDK,
but an old platform release might not take it. Therefore technically
speaking this is a breaking change. However we made a change of similar
nature in I66a762aa1734cf46cfc07f0570d535da38ddf1f7, and things didn't
blow up, so this should be fine.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/998667
Original-Revision: 3d49095e085dfcb9dd5fab8f5d488dbd54f4f7bc
GitOrigin-RevId: 1dbf4890647a1798a62bed24e9a3b9dc44b18f80
Change-Id: If894ce54789e54ce17afdc1626f5a9898ed1f2d4