[roll] Roll fuchsia [kernel][sched] Introduce sched::RunQueue
This change is the first commit to the new scheduler library, libsched.
It aims to refound the existing scheduler logic in a more encapsulated,
testable way. This will also ease the way to the introduction of
'hybrid scheduling'. In this change, we introduce one of the primary
scheduling abstractions, the "run queue", which manages the order in
which threads should be scheduled.

The thread abstraction introduced here is minimal, representing
'deadline' threads alone; support for 'fair' threads will be added in
later patches.

There are a few minor, intentional deviations here from how the related
scheduler state is expressed in the kernel today. In particular, instead
of that state being accessible via a `scheduler_state()` accessor as it
is with kernel thread class today, we make this part of a ThreadBase
base class. Several reasons for this:
* It is not out of place here, as this indeed concerns state of the
  thread itself and it is conceptually natural to access it off of a
  thread abstraction directly (e.g., `thread.capacity()`). Moreover, we
  introduce methods that are more naturally regarded and read as actions
  *on* the thread (e.g., thread.Reactivate()).
* Otherwise, we end up with a fair amount of 'duplicated' names in
  having both `Thread foo` and `State foo_state` (or `Thread foo_thread`
  and `State foo` in scope). This also adds more clutter and impacts
  readability a bit.
* It makes for cleaner code, inside of the library and out: in the
  library it's easier to assert that a given thread class inherits from
  the base, rather than, say, conforms to a particular interface;
  outside the library, easier for users to merely inherit from the
  subclass instead of recapitalting some desired interface.

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006654
Original-Revision: 9de46981679666cfbf18871da44d8db7b8220e0d
GitOrigin-RevId: dc1035ef164b413779397f47c46dcfc6606b7502
Change-Id: I86c12e7b8d0e2fe27d76a8217ac923fc82635f62