[roll] Roll fuchsia [phys] Ensure physboot-elf-kernel-test has a large kernel package
As an interim measure, physload reuses the space previously occupied by
the kernel package (i.e., the STORAGE_KERNEL ZBI item) for the hand-off
to the kernel. With profile instrumentation, this hand-off further
encodes profiles which can take up a fair amount of space. When the
kernel package is compressed, it is too small to fit a hand-off with
profiles, so we reserve more space in there for now by including the
kernel proper.

Original-Fixed: 325513615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011132
Original-Revision: 8ecbb4427982ae306999828c29140d39a2146865
GitOrigin-RevId: b0f8373c80a0050831cbce04b36ae67b788b24ac
Change-Id: Ia0003f7762c8456e8856fcedb7ebb53d68d9ebaf