[roll] Roll fuchsia [netstack3] Support MTU updates on Pure IP devices
Because the MTU can now be updated, it is dynamic state that needs to be
held behind a lock, with a corresponding lock level. While here, encode
accesses to the MTU through a `PureIpDeviceStateContext` trait which
will make it easier in the future to 1) add more state to Pure IP
devices, and 2) write test mocks.

Original-Bug: b/42051633
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007962
Original-Revision: 5192187267415773c4caee38c84b85de041f07be
GitOrigin-RevId: 11e9ffa9a80e01a735c1c45e51c53a1829102c6a
Change-Id: Ifce127ded143b4e5a7b66246edb0477d794ccd35