[roll] Roll fuchsia [assembly] Introduce set of experimental AIBs
An experimental AIB is a platform feature that is currently being
development, and we may want to test on a product before doing a
rigorous review and condoning it for production use.

These AIBs cannot be in user builds, and will not show up in userdebug
goldens, resulting in scrutiny failing for any user/userdebug product
that attempts to include them.

Additionally, moves adb_support into the experimental set and allows on
userdebug builds that don't run scrutiny.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971972
Original-Revision: d76277c045324472322ccc7693080718e52e63d2
GitOrigin-RevId: ae076e5de5a5dbc568f57b3ec3552ee6898470c4
Change-Id: I25a7af7aa1d08158d9e5dd5df76a2bd3e8ba3e03