[roll] Roll fuchsia [time][test] Removes unneeded expose stanzas
After some swapping the context back in, it seems that these
exposes are a remnant of something I tried to do, which failed,
and I then never cleaned up the remnants of.

In turn, this means that `test.time.realm.RealmFactory` is
the only source of in-realm protocols that should be used.

Removing the stanzas and comment now, to reduce future
confusion potential.

Original-Bug: b/327476961
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009326
Original-Revision: e3ed1dab4230c07fc3ad52b047de9a2d0cc4d6dc
GitOrigin-RevId: 39cb8634456d84a85282af392b8722aefb277212
Change-Id: Ied85eb0710f5551a0fbbb13458173d9c5367d15f