[roll] Roll fuchsia [netstack3] Move TCP to new TimerContext
Introduce CoreTimerContext in the base crate, which provides the new
timer-transformation mechanism that ties in with the new timer context.

Use the fact that we now need a type to control timers to provide
type-safety around the fact that only connected TCP sockets have timers,
which is a nice side-effect.

While I'm here, we've identified a race with tw reuse that can happen
between the reuse and timewait state timers firing. Relax the timer
handler and timer cancelation to deal with that.

Original-Bug: 42083407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009218
Original-Revision: 4c9b443b2aed103da5d8306c39987d81f17e9fe1
GitOrigin-RevId: f1d4d01816296b791ab3a1301d734282497c8329
Change-Id: I5bb8043fb0823d9982c27b9b23134eff0efacc36