[roll] Roll fuchsia [system-updater] Try the other slot when reading an asset errors
When checking to see if the non-current configuration already has the
correct asset fails with an error, system-updater used to fall back to
resolving the image package. Now it will first try to see if the
current configuration has the correct asset, the same as if checking
the non-current configuration failed just because the contents were
different.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010814
Original-Revision: 1b75894ea2add2334f04722265169182716bf193
GitOrigin-RevId: 0d9ba0fa21c03798039d0bcece95b349b5a5e800
Change-Id: Ie5ffbce927c23c34d038a53e3b87738b7a1b5acf