[roll] Roll fuchsia [kernel] do not sort text/data/bss sections by name
Keep sorting by alignment but don't sort by name after alignment.

Sorting by name is not particularly productive since most of the symbols
in the system are C++ name mangled, so the net result is largely sorting
by the length of the first mangled field, and even then it's an alphabetic
interpretation of a numeric field in the mangling.

This keeps symbols roughly in the same order they are in the particular
object files so in general keeps related code near each other in the
in the binary. Also makes it a little easier to read the disassembly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009328
Original-Revision: d565c8ff36bfe911a6279c269b6a4d31203df4ab
GitOrigin-RevId: 3e465ff4b854d0c8a3e4cd7e3ef007e153f3f992
Change-Id: I5db430075708cfcb0060a04e8cbb87dd319a5951