[roll] Roll fuchsia [check-licenses] Support landing fuchsia-infra-bazel-rules.
The fuchsia-infra-bazel-rules repo does not come with its own
README.fuchsia file, so one needs to be created for it in fuchsia.git.

The readme file can't be added to fuchsia.git in the same CL as the
project itself, since they exist in two different repositories.

This CL puts a temporary skip for the new project. It will be removed
shortly, after the project lands and a README.fuchsia file is
created in the //third_party/fuchsia-infra-bazel-rules folder.

Original-Bug: b/327010755
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011933
Original-Revision: ffd475d87bf2063ff046d271d071ab0ca86f25ee
GitOrigin-RevId: 46fb0c1575c7efbe3e8b3dd7bad6dae3c1becc93
Change-Id: Ia1affbc147c8f2b072b0f9111247d221002ae64a