[roll] Roll fuchsia [debugger] Add recursive component filters
Recursive filters will attach to all child components spawned under the
moniker of the given filter. The filter can be any variation of
component filters, for either component URLs or monikers. When the
filter is matched, the full moniker will be known by DebugAgent and a
special prefix filter installed (but not visible to users) will match
child components that are started.

This also works for attaching to components that do not have an ELF
program block, which can be the case for tests that are rooted with a
routing component.

Original-Bug: b/329916323
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011476
Original-Revision: 4281e2b21f5793014e095cc335d6fa7649c58d62
GitOrigin-RevId: b151221d68b1fdc8d28c3c64f4ce588b81da97b2
Change-Id: I93b989d3ca6d4205cfc3a3ec02aebddef45f7f4f