[roll] Roll fuchsia [bt][virtual] Migrate LoopbackDevice Driver from Banjo to FIDL
RootCanal and PTS-Bot use the new bt-host component to connect to the
LoopbackDevice driver using FIDL. This change updates the
LoopbackDevice to use the FIDL server end of the HCI protocol instead
of the Banjo version.

This is also part of the larger effort to migrate drivers to DFv2 and
move bt-host out of Fuchsia.

Original-Bug: b/329506134, b/303116559
Test:
- fx test bt-hci-virtual-tests
- Manual test with RootCanal, using ffx bluetooth subtool to start
  Pandora and send HCI packet between RootCanal and bt-host

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010218
Original-Revision: 08a3b4f782d0ca10b722ae3876e19d06c73bc797
GitOrigin-RevId: 9d688b713da06b3e6d35f21e2acd3269b4ce6876
Change-Id: I84d63e1054b9525ebbb8a5ccb9ee2a994864d67b