[roll] Roll fuchsia [starnix][selinux] Use policy-defined Ids in SecurityContext fields
Replace strings in SecurityContext field values with policy-defined
(numeric) identifiers.  This removes the need to map from strings
to Ids when looking up rules, and ensures that if a SID can be
resolved to a Context then the fields are guaranteed to have values
actually defined by the policy.

Consistency of fields e.g. range levels, whether the role is
valid for the user, will be landed separately.

Original-Bug: 328375715, 319232900, 330677360
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011912
Original-Revision: 05d36a2693d06cac5e7687bcbafa56b642c14119
GitOrigin-RevId: d64c415e6e89e32ff04a9670a3e093fddcb2b542
Change-Id: Ie85376450c737a98a19999b7297f8bf199bd7684