[roll] Roll fuchsia [doc] ffx component start doesn't update package
The component start process attempts to resolve the component
if not already resolved. However, if the component is already
resolved, it is not re-resolved again. Thus one cannot update
a component just by starting it. This incorrect documentation
is spotted in chat threads so we correct it.

The original table also doesn't make sense that you can update
the package without updating the manifest. The manifest is
contained in the package.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011238
Original-Revision: 3745e5a34d8c5cdd87b63d784d3d99ea854bfbf1
GitOrigin-RevId: 460e1df92fb929aa3bf0c8394c21e0576b1695da
Change-Id: Ib6c0522f2c62ee84502b4f29de3f9ce7ab149dd2