[roll] Roll fuchsia [bazel_sdk] Drop ABI with subpackages_to_flatten.
A subsequent change will forbid including the ABI stamp file in a
package manifest, since it could possibly conflict with the --api-level
specified on the CLI. This clears the way for that.

Also, adds a note to subpackages_to_flatten that it's a temporary
workaround.

Original-Bug: 330189874
Original-Bug: 326095332
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009256
Original-Revision: f0aa8b4995080c7764d805e3971c4251580bdd09
GitOrigin-RevId: 12bcaa6e80b3d29e0c65caf75ae920c6ad0e8f56
Change-Id: Ic89680fe3801f9bc02f3471537d0dc892e9ad85c