[roll] Roll fuchsia [fshost] Add 2 minute ramdisk client timeout to integration tests.
Some of these tests are flaking with timeouts that seem to be nearby
this call, which has historically been a source of slowness. Currently
the timeouts don't give any stack information, but it would be good to
be able to implicate what we're actually hanging on.

Original-Bug: b/326397893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011220
Original-Revision: 7be1b91ff1595f063691e345bc0e392d06dc1d4f
GitOrigin-RevId: ca7f5b7ff41df705556b18eaa295b476a91c58fd
Change-Id: I6a7c6e00e257aa48f3f14a80be30bc804129397b