[roll] Roll fuchsia [ramdisk] Set dead_ on DdkAdd failures.
Failures to call DdkAdd currently shutdown by calling DdkRelease on the
device where it will wait for the worker thread to join. The worker
thread will only complete when dead_ is set which is done in DdkUnbind
which will not be called through the failure path for DdkAdd.

Also logs failures to DdkAdd to make it clear what the failure is.

Original-Bug: b/326397893
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013634
Original-Revision: b8ff03d21394a2d807422cd517b2943805e66c0e
GitOrigin-RevId: ea6de4eb4f99dad9369460638a266af7dbce2cfa
Change-Id: I81df4e1b83bec70e36a2de474c5b25b4abaf2074