[roll] Roll fuchsia [ld] Add partial-page-bss test case
The existing modules don't explicitly exercise the partial-page
bss zeroing code paths.  It turned out they were being tested
only in coverage and profile builds where the test modules happen
to come out that way.  This adds a test tailor-mode to ensure it
will exercise those code paths in the loading implementations.
This is both the {Vmar,Mmap}Loader, which may already be tested
for these things, but also SegmentWithVmo::AlignSegments, which
was not tested in all paths in the elfldltl tests but is used in
the remote cases.  The new test exercises the missed paths there.

Original-Bug: 326524302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012253
Original-Revision: 1cac331f39dfad811cadf66ddb9a81c093703aca
GitOrigin-RevId: 7c02deab918858cf160722457dfdb3a79cad9d1b
Change-Id: I2bb56aa71fe94a2f5405dd772449f052a01e0b1a