[roll] Roll fuchsia [idk] Allow building for unsupported API level.
I tried to retire API level 14, and I found that doing so broke
core.x64-sdk_source_sets_and_shlibs-api14-build_only, because we make
sure we don't try to build the IDK targeting an unsupported API level,
but that's precisely what that builder does.

Infra's view of which API levels are supported is unfortunately
necessarily out of sync with the reality in fuchsia.git. The right thing
to do is to delete those API-level-specific builders and rely on
subbuilds instead, but for now we can just remove the assertion to
unblock retiring unused API levels.

Original-Bug: 330709069
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009262
Original-Revision: ddea9daf2b6a36950160812f26965b5efb601020
GitOrigin-RevId: d0e9dcdbf5ccd058bd24af450d9b3165dc1b5f7b
Change-Id: I423a618134e50c6ea4855e7705d5c5c12c410101