[roll] Roll fuchsia [pkg][blobfs-client] Don't log in open when blob is missing
This log is emitted when the retained index is set, which is generally
expected and can be misleading. The log was more useful when support
for the new FIDL api was being added. The logged status is still
available in the epitaph.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013153
Original-Revision: b9098434fdd00cafd9228a924a2b9ca4b6f76766
GitOrigin-RevId: 5d896f2c3335f1782d8bfb2d86360cccc3bbca9d
Change-Id: Id48ba059a1c6cc25eb4a987736a60ac6d9029480