[roll] Roll fuchsia [vfs] Add array-backed simple files
This is a way to create a simple file that works on both host and target
(i.e. with or without VMOs).

On host we present the `read_only` function in the top of the file
module, and on target we re-export the vmo-based `read_only` in the same
position.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012880
Original-Revision: 1c3534841906f9d54d65e8be99c558cc73815676
GitOrigin-RevId: a9118e73ce91f48ebe79db21e3b5ba01d25e980e
Change-Id: I6175b107def12fb303a9515166fb3a1cd40a15a4