[roll] Roll fuchsia [aml-sdmmc] Minor refactoring in preparation for Power integration.
No behavioural changes are intended in this CL.

This change decouples power awareness from whether the incoming request
is of Banjo or FIDL. This is useful because Banjo and FIDL have
different request processing models. For example, this refactoring makes
it easy to delay a FIDL request and serve (complete) it at a later time
using the Async completer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011472
Original-Revision: 1dad9c0c10620cea753af99544015c4be337962b
GitOrigin-RevId: 71abf60e842237fe74b754523f16fb32092cb665
Change-Id: I694874bba4750ca65f7e5a0be59591b34409b016