[roll] Roll fuchsia [ffx] Rename to clarify role of target specifier
The name `get_default_target()` was misleading, as it didn't get
a target (it got a specifier) and it didn't get a default (e.g.
providing "-t <spec>" on the command line goes through the
same code).

To that end, the function, some variables, and some comments have
been updated to clarify the role of the function.

Original-Bug: b/320519654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009226
Original-Revision: 8f911fca5991063a074ce21c4053d7614b7a4af3
GitOrigin-RevId: 06094e1d0a687e3032e7ab7f4978693ff08c559f
Change-Id: I15c5121ba7e9fca928b2c5fc76c05cd9585e84b0