[roll] Roll fuchsia [fidl] Optimize the client waker
On Fuchsia, we can read the channel when we are woken and then only open
the task that needs to be woken rather than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010794
Original-Revision: 9fe20d92e5e5b6ab4724ac136ac65ebaffb002bd
GitOrigin-RevId: beb7c8c884b828537bee5b953f329bee921996a9
Change-Id: I9b7dbe3c6f6d2c093b58a1277a9f1fa7928f011f