[roll] Roll fuchsia [bt] Add the credit-based flow control TxEngine
Add support for a TxEngine that implements credit-based flow control.
The flow control scheme tracks a number of credits that the peer has
provided, and allows sending a frame only if there are sufficient
credits to send the frame.

Currently this only supports LE credit-based flow control, but the
implementation is designed to be able to be easily modifiable to support
enhanced CBFC later.

Test: fx test bt-host-l2cap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002193
Original-Revision: 2b4d1b1c21a8805060716ff7be8d799ef96c1939
GitOrigin-RevId: 13d93c9b93f363ae9806cdbcf7ca223ac5436bc9
Change-Id: Ibea32008743cbbb8f48c34c40cbc33555cc42f00