[roll] Roll fuchsia [heapdump][instrumentation] Do not dynamically link libunwind.so.1
When libheapdump_instrumentation.so is used by SDK clients,
libunwind.so.1 may not exist or have a different ABI in the client's
toolchain.

We originally used to link it statically using the
//build/config/fuchsia:static_cpp_standard_library config option, but
fxrev.dev/919272 removed it to avoid a warning, introducing the runtime
dependency on libunwind.so.1 by mistake.

This CL restores the old correct behavior, but in a way that does not
trigger the warning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013656
Original-Revision: d32d25ab25286365c806d4e5e1ffcaca18d8f153
GitOrigin-RevId: cc1bd1442aeac06df2a8a56fd3e09e333a2e5614
Change-Id: Ie266934f01368e990b9b25855ae34171221e47a3