[roll] Roll fuchsia [bt-hci] Remove extra init dispatcher
Removes the extra dispatcher that was being used in the executor used
for initialization. This is no longer needed and causes unnecessary
complexity in the code. This also fixes the flake that was happening due
to the extra dispatcher.

Original-Bug: 329168584
Multiply: bt-hci-broadcom-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012886
Original-Revision: 464dd0094b64ed01d44a410d20fb7f7df308ad16
GitOrigin-RevId: 4e65e2353a7032e45fbde508d70399597d88df54
Change-Id: I9431c75404c004b397509b18e1d9214f8b5d282b