[roll] Roll fuchsia [debugger] Notify when installing recursive filter
Adds a new notification to be sent from the Agent to the client,
notifying that a ComponentDiscovered event was received by the
ComponentManager and a recursive filter has been applied to match other
components being launched in this realm.

Currently, only the session cares about this notification, so there is
no modification to the SessionObserver interface, but can be easily
expanded to include an observer notification.

Original-Fixed: b/330571289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013624
Original-Revision: fd6bd116451f57965b4b4ed47d4c3d7d87b6861c
GitOrigin-RevId: 97c57e3b1571e82c9bb9653dae80da7b3a25ed95
Change-Id: I7d034c61033c8baffd2ce577cf4250079d69e2ff