[roll] Roll fuchsia [sdk][vfs] Fix PseudoFile write handler
If a write handler is not specified, the file should be read only. The
write handler was unconditionally initialized even if one was not
specified, however. This would lead to a nullptr dereference if the file
was opened as writable.

Test: fx test vfs_cpp_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014612
Original-Revision: f44378ca36792ee431da02fb578a84cb98fc23e7
GitOrigin-RevId: 508a0b4715dda43bac70fbd32f5e68628e141892
Change-Id: Ia61c1b81ceac18b74ec1969216824c1b4f720271